==> 1.1 A.R <==
# Loading useful packages ----
library(pacman)
p_load(dplyr, ggplot2, quanteda, readtext, knitr, stringr, tidyr, tidytext, beepr)

# library(tm) text mining package
# library(xlsx) for reading, writing, and formatting Excel files.
# library(ggpage) for arranging ggplot2 plots across multiple pages.

# Functions used from the stringr package:

# Retrieving and reading the documents to be used ----
Corpus <- readtext("./AR.TXT/*.pdf")

# load("Corpus.Rda")

# Creating the Stopwords (ENG) object ----
stop <- tibble(words = stopwords::data_stopwords_stopwordsiso$en)

# Changing the name of variable for doc_id
Corpus <- Corpus %>%
  mutate(Year = str_match(string = doc_id, "\\d{4}"),
         Year = as.numeric(Year)) %>%
  select(-doc_id)


# Cleaning the most salient things to erase, unless punctuation ----
# data.clean is the file before the unnest function (one text per row), with the 1st layer of
# cleaning (words wrongly written) & lower cases

## Cleaning and creating the data with NO punctuation ----
data.clean.no.punct <- Corpus  %>%
  mutate(text = str_replace_all(text, "\\n", " "), # every jump of line (in the text) is replaced by a space
         text = str_remove_all(text, "[:punct:]"), # removes every punctuation character
         text = str_remove_all(text, "[:digit:]"), # removes digits
         text = str_remove_all(text, "_+"), # removes underscores
         text = str_remove_all(text, "(\\$|\\~|c~|\\^|\\£|\\^|<>|<|>|¯|>>|⇒|→|>oo|\\boo\\b|°|±|™|\\+|\\-|\\¥|\\■|\\|)\\w?"), # removes specific symbols and patterns
         text = tolower(text), # converts uppercases to lowercases
         text = str_replace_all(text, "wellbeing", "well-being"),
         text = str_remove_all(text,"\\bwe the rockefeller foundation\\b"),
         text = str_remove_all(text,"wwwrockfoundorg"),
         text = str_remove_all(text,"wwwrockefeller"),
         text = str_replace_all(text, "erice(s)?\\b", "ence\\1"),
         text = str_replace_all(text, "univer\\s*(.|usa)?\\s*sity", "university"),
         text = str_replace_all(text, "\\b(organi|moderni)\\s+(.|usa)?\\s+zation", "\\1zation"),
         text = str_replace_all(text, "appropria(\\s|\\t|\\n)+tions", "appropriations"),
         text = str_replace_all(text, "\\b(ac|ag|aq)\\spriations", "\\1priations"),
         text = str_remove_all(text, "\\brockefeller foundation\\b"),
         text = str_replace_all(text, "\\b(adjust|equip)\\s\\b[pjo]\\b", "\\1"),
         text = str_replace_all(text, "nonallgn", "non alligned"),
         text = str_replace_all(text, "(govern|depart|treat|equip|develop|require|move|impove|endow|employ|install|agree|
         establish|disburse|adjust|supple|experi|settle|pay|invest|ele|achieve|commit|appoint|docu|arrange|assign)\\sment", "\\1ment"),
         text = str_replace_all(text, "\\b([a-z]{1,12})\\s\\bment(s?)\\b", "\\1ment\\2"),
         text = str_replace_all(text, "devel\\sopment", "development"),
         text = str_replace_all(text, "de\\svelopment", "development"),
         text = str_replace_all(text, " depart\\sments", "departments"),
         text = str_replace_all(text, "\\b([a-z]{0,9})\\s\\bcation(s?)\\b", "\\1cation\\2"),
         text = str_replace_all(text, "telecommun", "telecommuni"),
         text = str_replace_all(text, "\\b([a-z]{1,10})\\slation(s?)", "\\1lation\\2"),
         text = str_replace_all(text, "foun\\sdation", "foundation"),
         text = str_replace_all(text, "accomo\\sdation", "accomodation"),
         text = str_replace_all(text, "comis\\sj", "comis"),
         text = str_replace_all(text, "([a-z]{2,9})\\ssion(s?)", "\\1ssion\\2"),
         text = str_replace_all(text, "([a-z]{2,12})\\sion(s?)", "\\1sion\\2"),
         text = str_replace_all(text, "foun\\s(a|i)", "foun"),
         text = str_replace_all(text, "con\\struction", "construction"),
         text = str_replace_all(text, "oo struction", "construction"),
         text = str_replace_all(text, "indebted\\s(a|k)", "indebted"),
         text = str_replace_all(text, "([a-z]{0,11})\\s\\bness(e?s?)\\b", "\\1ness\\2"),
         text = str_replace_all(text, "feebleminded ness", "feeblemindedness"),
         text = str_replace_all(text, "\\bcom\\sq", "com"),
         text = str_replace_all(text, "\\bcom\\smunit(y|ies)", "communit\\1"),
         text = str_replace_all(text, "im\\smunity", "immunity"),
         text = str_replace_all(text, "corpo\\s~", "corpo"),
         text = str_replace_all(text, "\\b([a-z]{0,9})\\s\\bration(s?)\\b", "\\1ration\\2"),
         text = str_replace_all(text, "\\badminis\\s>\\sj", "adminis"),
         text = str_replace_all(text, "\\b([a-z]{2,9})\\s\\btration(s?)\\b", "\\1tration\\2"),
         text = str_replace_all(text, "([a-z]{2,10})\\s\\btional(s?)\\b", "\\1tional\\2"),
         text = str_replace_all(text, "\\b([a-z]{2,10})\\s\\btural(s?)\\b", "\\1tural\\2"),
         text = str_replace_all(text, "\\b([a-z]{1,11})\\s\\bation(s?)\\b", "\\1ation\\2"),
         text = str_replace_all(text, "\\b([a-z]{1,10})\\s\\btation(s?)\\b", "\\1tation\\2"),
         text = str_replace_all(text, "\\b([a-z]{2,12})\\s\\btion(s?)\\b", "\\1tion\\2"),
         text = str_replace_all(text, "\\b([a-z]{1,13})\\s\\bit(y|ies)\\b", "\\1it\\2"),
         text = str_replace_all(text, "([a-z]{1,12})\\s\\bence(s?)\\b", "\\1ence\\2"),
         text = str_replace_all(text, "\\b([a-z]{2,11})\\s\\btive(s?)\\b", "\\1tive\\2"),
         text = str_replace_all(text, "attenu ated", "attenuated"),
         text = str_replace_all(text, "\\boper ated", "operated"),
         text = str_replace_all(text, "appropri ated", "appropriated"),
         text = str_replace_all(text, "initi ated", "initiated"),
         text = str_replace_all(text, "evalu ated", "evaluated"),
         text = str_replace_all(text, "cre ated", "created"),
         text = str_replace_all(text, "associ ated", "associated"),
         text = str_replace_all(text, "situ ated", "situated"),
         text = str_replace_all(text, "affili oo ated", "affiliated"),
         text = str_replace_all(text, "cooper ated", "cooperated"),
         text = str_replace_all(text, "unaffili", "unaffiliated"),
         text = str_replace_all(text, "exagger ated", "exaggerated"),
         text = str_replace_all(text, "jniver ~", "univers"),
         text = str_replace_all(text, "ca lamities", "calamities"),
         text = str_replace_all(text, "accaountantfacil", "accaountant facil "),
         text = str_replace_all(text, "popu atlon", "population"),
         text = str_replace_all(text, "expendi tures", "expenditures"),
         text = str_replace_all(text, "corpoi", "corpor"),
         text = str_replace_all(text, "oper >", "oper"),
         text = str_replace_all(text, "collaboi", "collabor"),
         text = str_replace_all(text, "pos sible", "possible"),
         text = str_replace_all(text, "ap propri", "appropri"),
         text = str_replace_all(text, "populationappropri", "population appropri"),
         text = str_replace_all(text, "admin istration", "administration"),
         text = str_replace_all(text, "expendi ", "expenditures"),
         text = str_replace_all(text, "cd room", "cd-room"),
         text = str_replace_all(text, "nce", "rice"),
         text = str_replace_all(text, "viricent", "vincent"),
         text = str_replace_all(text, "organizuion", "organization"),
         text = str_replace_all(text, "autnoma", "autonoma"),
         text = str_replace_all(text, "ommunities", "communities"),
         text = str_replace_all(text, "ommunity", "community"),
         text = str_replace_all(text, "insurarice", "insurance"),
         text = str_replace_all(text, "ommunities", "communities"),
         text = str_replace_all(text, "californiaa", "california"),
         text = str_replace_all(text, "cccommunity", "community"),
         text = str_replace_all(text, "ccommunity", "community"),
         text = str_replace_all(text, "cccommunities", "communities"),
         text = str_replace_all(text, "ccommunities", "communities"),
         text = str_replace_all(text, "certifi cates", "certificates"),
         text = str_replace_all(text, "cer tificates", "certificates"),
         text = str_replace_all(text, "theadvancement", "the advancement"),
         text = str_replace_all(text, "thedevelopment", "the development"),
         text = str_replace_all(text, "thedivision", "the division"),
         text = str_replace_all(text, "thedirection", "the direction"),
         text = str_replace_all(text, "thegovernment", "the government"),
         text = str_replace_all(text, "thecommission", "the commission"),
         text = str_replace_all(text, "theestablishment", "the establishment"),
         text = str_replace_all(text, "thepopulation", "the population"),
         text = str_replace_all(text, "theimprovement", "the improvement"),
         text = str_replace_all(text, "thereproduction", "the reproduction"),
         text = str_replace_all(text, "theassociation", "the association"),
         text = str_replace_all(text, "thetreatment", "the treatment"),
         text = str_replace_all(text, "thepreparation", "the preparation"),
         text = str_replace_all(text, "theadvancement", "the advancement"),
         text = str_replace_all(text, "theorganization", "the organization"),
         text = str_replace_all(text, "thecorporation", "the corporation"),
         text = str_replace_all(text, "thedistribution", "the distribution"),
         text = str_replace_all(text, "theintroduction", "the introduction"),
         text = str_replace_all(text, "thecreation", "the creation"),
         text = str_replace_all(text, "thecooperation", "the cooperation"),
         text = str_replace_all(text, "theinfection", "the infection"),
         text = str_replace_all(text, "theconstruction", "the construction"),
         text = str_replace_all(text, "theapplication", "the application"),
         text = str_replace_all(text, "thesolution", "the solution"),
         text = str_replace_all(text, "theinvestigation", "the investigation"),
         text = str_replace_all(text, "thecompletion", "the completion"),
         text = str_replace_all(text, "thestation", "the station"),
         text = str_replace_all(text, "thepromotion", "the promotion"),
         text = str_replace_all(text, "theuniversity", "the university"),
         text = str_replace_all(text, "thefoundation", "the foundation"),
         text = str_replace_all(text, "thedepartment", "the department"),
         text = str_replace_all(text, "thefoundations", "the foundations"),
         text = str_replace_all(text, "thehumanities", "the humanities"),
         text = str_replace_all(text, "thecity", "the city"),
         text = str_replace_all(text, "theappropriations", "the appropriations"),
         text = str_replace_all(text, "therockefeller", "the rockefeller"),
         text = str_replace_all(text, "theeditors", "the editors"),
         text = str_replace_all(text, "thegrants", "the grants"),
         text = str_replace_all(text, "theblood", "the blood"),
         text = str_replace_all(text, "thecommittee", "the committee"),
         text = str_replace_all(text, "thebureau", "the bureau"),
         text = str_replace_all(text, "theimportant", "the important"),
         text = str_replace_all(text, "theressources", "the ressources"),
         text = str_replace_all(text, "witheressources", "the ressources"),
         text = str_replace_all(text, "theinternational", "the international"),
         text = str_replace_all(text, "thepractice", "the practice"),
         text = str_replace_all(text, "theproject", "the project"),
         text = str_replace_all(text, "theworld", "the world"),
         text = str_replace_all(text, "theinternal", "the internal"),
         text = str_replace_all(text, "thecouncil", "the council"),
         text = str_replace_all(text, "theeconomic", "the economic"),
         text = str_replace_all(text, "thenational", "the national"),
         text = str_replace_all(text, "themedical", "the medical"),
         text = str_replace_all(text, "thework", "the work"),
         text = str_replace_all(text, "thesocial", "the social"),
         text = str_replace_all(text, "thecalifornia", "the california"),
         text = str_replace_all(text, "thestate", "the state"),
         text = str_replace_all(text, "theothers", "the others"),
         text = str_replace_all(text, "andequipment", "and equipment"),
         text = str_replace_all(text, "andtreatment", "and treatment"),
         text = str_replace_all(text, "andgovernment", "and government"),
         text = str_replace_all(text, "anddistribution", "and distribution"),
         text = str_replace_all(text, "andfixedequipment", "and fixed equipement"),
         text = str_replace_all(text, "anddevelopment", "and development"),
         text = str_replace_all(text, "andalso", "and also"),
         text = str_replace_all(text, "andcommodities", "and commodities"),
         text = str_replace_all(text, "andeconomic", "and economic"),
         text = str_replace_all(text, "andwomen", "and women"),
         text = str_replace_all(text, "andrelationships", "and relationships"),
         text = str_replace_all(text, "andscientists", "and scientists"),
         text = str_replace_all(text, "andpossibilities", "and possibilities"),
         text = str_replace_all(text, "andexpanded", "and expanded"),
         text = str_replace_all(text, "andaccordingly", "and accordingly"),
         text = str_replace_all(text, "ofinfection", "of infection"),
         text = str_replace_all(text, "ofdevelopment", "of development"),
         text = str_replace_all(text, "offoundation", "of foundation"),
         text = str_replace_all(text, "ofpopulation", "of population"),
         text = str_replace_all(text, "ofequipment", "of equipment"),
         text = str_replace_all(text, "oftreatment", "of treatment"),
         text = str_replace_all(text, "ofnutrition", "of nutrition"),
         text = str_replace_all(text, "ofinvestigations", "of investigations"),
         text = str_replace_all(text, "ofinformation", "of informtion"),
         text = str_replace_all(text, "ofrecognized", "of recognised"),
         text = str_replace_all(text, "ofmedical", "of medical"),
         text = str_replace_all(text, "ofeducation", "of education"),
         text = str_replace_all(text, "ofminorityyouth", "of minority youth"),
         text = str_replace_all(text, "ofhealth", "of health"),
         text = str_replace_all(text, "ofdistinguished", "of distinguished"),
         text = str_replace_all(text, "ofsanitary", "of sanitary"),
         text = str_replace_all(text, "ofresearch", "of research"),
         text = str_replace_all(text, "overthe", "over the"),
         text = str_replace_all(text, "withtheressources", "with the ressources"),
         text = str_replace_all(text, "withmore", "with more"),
         text = str_replace_all(text, "withfoundation", "with foundation"),
         text = str_replace_all(text, "toaccomplish", "to accomplish"),
         text = str_replace_all(text, "tooperate", "to operate"),
         text = str_replace_all(text, "toinequality", "to inequality"),
         text = str_replace_all(text, "toincrease", "to increase"),
         text = str_replace_all(text, "tothe", "to the"),
         text = str_replace_all(text, "toassist", "to assist"),
         text = str_replace_all(text, "aneducation", "an education"),
         text = str_replace_all(text, "anappropriation", "an appropriation"),
         text = str_replace_all(text, "theparticipating", "the participating"),
         text = str_replace_all(text, "afoundation", "a foundation"),
         text = str_replace_all(text, "asingle", "a single"),
         text = str_replace_all(text, "whetherthey", "wether they"),
         text = str_replace_all(text, "foodproduction", "food production"),
         text = str_replace_all(text, "incollaboration", "in collaboration"),
         text = str_replace_all(text, "incooperation", "in cooperation"),
         text = str_replace_all(text, "inaddition", "in addition"),
         text = str_replace_all(text, "infieldof", "in field of"),
         text = str_replace_all(text, "fordevelopment", "for development"),
         text = str_replace_all(text, "forcollaborative", "for collaborative"),
         text = str_replace_all(text, "foruse", "for use"),
         text = str_replace_all(text, "forthe", "for the"),
         text = str_replace_all(text, "healthdivision", "health division"),
         text = str_replace_all(text, "healthorganization" , "health organization"),
         text = str_replace_all(text, "healthdepartment", "health department"),
         text = str_replace_all(text, "healtheducation", "health education"),
         text = str_replace_all(text, "healthadministration", "health administration"),
         text = str_replace_all(text, "healthcommission", "health commission"),
         text = str_replace_all(text, "highereducation", "higher education"),
         text = str_replace_all(text, "medicaleducation", "medical education"),
         text = str_replace_all(text, "publicadministration", "public administration"),
         text = str_replace_all(text, "foundationfoundation", "foundation"),
         text = str_replace_all(text, "foundationtable", "foundation table"),
         text = str_replace_all(text, "foundationcontents", "foundation contents"),
         text = str_replace_all(text, "foundationgrants", "foundation grants"),
         text = str_replace_all(text, "foundationuniversity", "foundation university"),
         text = str_replace_all(text, "foundationscience", "foundation science"),
         text = str_replace_all(text, "foundationintroductionthe", "foundation
                                introduction the"),
         text = str_replace_all(text, "foundationsystematically", "foundation systematically"),
         text = str_replace_all(text, "foundationsupported", "foundation supported"),
         text = str_replace_all(text, "foundationgrant", "foundation grant"),
         text = str_replace_all(text, "foundationabout", "foundation about"),
         text = str_replace_all(text, "foundationin", "foundation in"),
         text = str_replace_all(text, "foundationadministered", "foundation administered"),
         text = str_replace_all(text, "foundationtion", "foundation"),
         text = str_replace_all(text, "foundationphotograph", "foundation photograph"),
         text = str_replace_all(text, "foundationon", "foundation on"),
         text = str_replace_all(text, "foundationbalarice", "foundation balance"),
         text = str_replace_all(text, "foundationopportunities", "foundation opportunities"),
         text = str_replace_all(text, "soilpollution", "soil pollution"),
         text = str_replace_all(text, "soilshokkaido", "soils hokkaido"),
         text = str_replace_all(text, "wuniversity", "university"),
         text = str_replace_all(text, "universitythe", "university the"),
         text = str_replace_all(text, "univesitythe", "university the"),
         text = str_replace_all(text, "umversidad", "universidad"),
         text = str_replace_all(text, "communitydevelopment", "community development"),
         text = str_replace_all(text, "hookworminfection", "hookworm infection"),
         text = str_replace_all(text, "economicdevelopment", "economic development"),
         text = str_replace_all(text, "reportfoundation", "report foundation"),
         text = str_replace_all(text, "nursingeducation", "nursing education"),
         text = str_replace_all(text, "educationagriculture", "education agriculture"),
         text = str_replace_all(text, "specialattention", "special attention"),
         text = str_replace_all(text, "sciencenutrition", "science nutrition"),
         text = str_replace_all(text, "ruralreconstruction", "rural reconstruction"),
         text = str_replace_all(text, "complementfixation", "complement fixation"),
         text = str_replace_all(text, "saltflotation", "salt flotation"),
         text = str_replace_all(text, "rockfellerfoundation", "rockfeller foundation"),
         text = str_replace_all(text, "perioduniversity", "period university"),
         text = str_replace_all(text, "grantsuniversity", "grants university"),
         text = str_replace_all(text, "internationalrelations", "international relations"),
         text = str_replace_all(text, "guniversity", "university"),
         text = str_replace_all(text, "columbiauniversity", "columbia university"),
         text = str_replace_all(text, "hssuniversity", "hss university"),
         text = str_replace_all(text, "paymentsuniversities", "payments universities"),
         text = str_replace_all(text, "forfiveyears", "for five years"),
         text = str_replace_all(text, "highyielding", "high yielding"),
         text = str_replace_all(text, "highquality", "high quality"),
         text = str_replace_all(text, "arthropodborne", "arthropod-borne"),
         text = str_replace_all(text, "nee", "rice"),
         text = str_replace_all(text, "hivaids", "hiv aids"),
         text = str_replace_all(text, "tne", "the"),
         text = str_replace_all(text, "latrines", "latrine"),
         text = str_replace_all(text, "otherexhibition", "other exhibition"),
         text = str_replace_all(text, "rockefellerfoundation", "rockefeller foundation"),
         text = str_replace_all(text, "harvarduniversity", "harvard university"),
         text = str_replace_all(text, "aboveuniversity", "above university"),
         text = str_replace_all(text, "generaleducation", "general education"),
         text = str_replace_all(text, "nonproliferation", "non-proliferation"),
         text = str_replace_all(text, "periodunivkhsity", "period university"),
         text = str_replace_all(text, "varietiestanzania", "varieties tanzania"),
         text = str_replace_all(text, "policieshuman", "policies human"),
         text = str_replace_all(text, "salaamtanzania", "salaam tanzania"),
         text = str_replace_all(text, "salaamfrom", "salaam from"),
         text = str_replace_all(text, "ascomprehensive", "as comprehensive"),
         text = str_replace_all(text, "characterizationbiotechnologyscnpps",
                                "characterization biotechnology scnpps"),
         text = str_replace_all(text, "associatepatricia", "associate patricia"),
         text = str_replace_all(text, "associatejohn", "associate john"),
         text = str_replace_all(text, "biologyjudge", "biology judge"),
         text = str_replace_all(text, "minoritygroup", "minority group"),
         text = str_replace_all(text, "europeanracial", "european racial"),
         text = str_replace_all(text, "mexicanamerican", "mexican american"),
         text = str_replace_all(text, "peruuniversity", "peru university"),
         text = str_replace_all(text, "japanhokkaido", "japan hokkaido"),
         text = str_replace_all(text, "japangraduate", "japan graduate"),
         text = str_replace_all(text, "japanhague", "japan hague"),
         text = str_replace_all(text, "japaneseunited", "japanese united"),
         text = str_replace_all(text, "colombiauniversity", "colombia university"),
         text = str_replace_all(text, "colombiaversity", "colombia university"),
         text = str_replace_all(text, "brazillaboratory", "brazil laboratory"),
         text = str_replace_all(text, "availableto", "available to"),
         text = str_replace_all(text, "fourfellowships", "four fellowships"),
         text = str_replace_all(text, "valuespririceton", "values princenton"),
         text = str_replace_all(text, "medicineuniversity", "medicine university"),
         text = str_replace_all(text, "facultyof", "faculty of"),
         text = str_replace_all(text, "veterinarymedicine", "veterinary medicine"),
         text = str_replace_all(text, "biologygenetics", "biologoy genetics"),
         text = str_replace_all(text, "developmentsustainable", "development sustainable"),
         text = str_replace_all(text, "projectsupport", "project support"),
         text = str_replace_all(text, "autnoma", "autonoma"),
         text = str_replace_all(text, "darice", "dance"),
         text = str_replace_all(text, "entrarice", "entrance"),
         text = str_replace_all(text, "substarices", "substances"),
         text = str_replace_all(text, "importarice", "importance"),
         text = str_replace_all(text, "coricentrating", "concentrating"),
         text = str_replace_all(text, "periodjapanhokkaido", "period japan hokkaido"),
         text = str_replace_all(text, "sociologydemography", "sociology demography"),
         text = str_replace_all(text, "laboratoryteaching", "laboratory teaching"),
         text = str_replace_all(text, "httpwwwhindustantimescomnewsfeedindiaindiacheersasmonsoonarriveshopesofbetterfarmoutputraisedarticleaspx",
                                "http www hindustan times com news feed india india cheersas monsoon arrives hopes of better farm out put raised article aspx"),
         text = str_replace_all(text, "resiliericestrategies", "resilience strategies"),
         text = str_replace_all(text, "broadlyshared", "broadly shared"),
         text = str_replace_all(text, "changewhile", "change while"),
         text = str_replace_all(text, "insurariceproducts", "insurance products"),
         text = str_replace_all(text, "themworldand", "them world an"),
         text = str_replace_all(text, "smallholder", "small holder"),
         text = str_replace_all(text, "climatechange", "climate change"),
         text = str_replace_all(text, "activitiesexhibition", "activities exhibition"),
         text = str_replace_all(text, "testimonyvernacular", "testimony vernacular"),
         text = str_replace_all(text, "collaborationand", "collaboration and"),
         text = str_replace_all(text, "providechroreographic", "provide choreographic"),
         text = str_replace_all(text, "anyonesexpextation", "anyones expectation"),
         text = str_replace_all(text, "specialscholarships", "special scholarships"),
         text = str_replace_all(text, "scholarshipassistarice", "scholarship assistance"),
         text = str_replace_all(text, "automaticdrips", "automatic drips"),
         text = str_replace_all(text, "studiesand", "studies and"),
         text = str_replace_all(text, "trainingbases", "training bases"),
         text = str_replace_all(text, "bloodforming", "blood forming"),
         text = str_replace_all(text, "basicbiological", "basic biological"),
         text = str_replace_all(text, "pathologicalthrough", "pathological through"),
         text = str_replace_all(text, "substariceswhich", "substances which"),
         text = str_replace_all(text, "arrangementshave", "arrangement shave"),
         text = str_replace_all(text, "statesintensive", "states intensive"),
         text = str_replace_all(text, "whichhave", "which have"),
         text = str_replace_all(text, "southernstates", "southern states"),
         text = str_replace_all(text, "morethat", "more that"),
         text = str_replace_all(text, "theirresponsibility", "their responsibility"),
         text = str_replace_all(text, "degreeof", "degree of"),
         text = str_replace_all(text, "entirerepublic", "entire republic"),
         text = str_replace_all(text, "effectivemethods", "effective methods"),
         text = str_replace_all(text, "averagedegree", "average degree"),
         text = str_replace_all(text, "governmentprivate", "government private"),
         text = str_replace_all(text, "betterqualified", "better qualified"),
         text = str_replace_all(text, "residualspraying", "residual spraying"),
         text = str_replace_all(text, "tworepresentative", "two representative"),
         text = str_replace_all(text, "worldwideusefulness", "worldwide usefullness"),
         text = str_replace_all(text, "favorabletime", "favorable time"),
         text = str_replace_all(text, "methodthe", "method the"),
         text = str_replace_all(text, "authoritieshowever", "authorities however"),
         text = str_replace_all(text, "decreasein", "decrease in"),
         text = str_replace_all(text, "morelimited", "more limited"),
         text = str_replace_all(text, "whichhas", "which has"),
         text = str_replace_all(text, "isan", "is an"),
         text = str_replace_all(text, "truethat", "true that"),
         text = str_replace_all(text, "decadebecomes", "decade becomes"),
         text = str_replace_all(text, "socialsanitary", "social sanitary"),
         text = str_replace_all(text, "theirdistribution", "their distribution"),
         text = str_replace_all(text, "housingfamily", "housing family"),
         text = str_replace_all(text, "problemswhich", "problems which"),
         text = str_replace_all(text, "byfor", "by for"),
         text = str_replace_all(text, "feasabilitypower", "feasability power"),
         text = str_replace_all(text, "prospectthought", "prospect thought"),
         text = str_replace_all(text, "inprogress", "in progress"),
         text = str_replace_all(text, "selfsufficiency", "self sufficiency"),
         text = str_replace_all(text, "providessupport", "provides support"),
         text = str_replace_all(text, "agegroups", "age groups"),
         text = str_replace_all(text, "proposedpersonal", "proposed personal"),
         text = str_replace_all(text, "scholarsphilanthropic", "scholars philanthropic"),
         text = str_replace_all(text, "noeconomic", "no economic"),
         text = str_replace_all(text, "journalismeducationberkeley",
                                "journalism education berkeley"),
         text = str_replace_all(text, "towardsits", "towards its"),
         text = str_replace_all(text, "numberof", "number of"),
         text = str_replace_all(text, "newspaperindustry", "newspaper industry economic"),
         text = str_replace_all(text, "robinsonfoundationnew",
                                "robinson foundation new"),
         text = str_replace_all(text, "fundadevelopment", "fund a development"),
         text = str_replace_all(text, "citizensexpanding", "citizens expanding"),
         text = str_replace_all(text, "businesscollaboratweprogramsaimedat",
                                "business collaborat we programs aimed at"),
         text = str_replace_all(text, "improvingyouth", "improving youth"),
         text = str_replace_all(text, "eemployment", "employment"),
         text = str_replace_all(text, "testedin", "tested in"),
         text = str_replace_all(text, "careeropportunities", "career opportunities"),
         text = str_replace_all(text, "technologyand", "technology and"),
         text = str_replace_all(text, "developmentsaffecting", "developments affecting"),
         text = str_replace_all(text, "givingsupport", "giving support"),
         text = str_replace_all(text, "sciencesmathematics", "sciences mathematics"),
         text = str_replace_all(text, "assistarice", "assistance"),
         text = str_replace_all(text, "iriception", "inception"),
         text = str_replace_all(text, "coricentrated", "concentrated"),
         text = str_replace_all(text, "coriception", "conception"),
         text = str_replace_all(text, "coricerned", "concerned"),
         text = str_replace_all(text, "sirice", "since"),
         text = str_replace_all(text, "coricentration", "concentration"),
         text = str_replace_all(text, "socialadvarice", "social advance"),
         text = str_replace_all(text, "comparativecommunity", "comparative community"),
         text = str_replace_all(text, "ministration", "administration"),
         text = str_replace_all(text, "appf priation", "appropriation"),
         text = str_replace_all(text, "appro priation", "appropriation"),
         text = str_replace_all(text, "priation", "appropriation"),
         text = str_replace_all(text, "internationalhealth", "international health"),
         text = str_replace_all(text, "publichealth", "public health"),
         text = str_replace_all(text, "interna tional", "international"),
         text = str_replace_all(text, "addi tional", "additional"),
         text = str_replace_all(text, "aroundadvarices", "around advances"),
         text = str_replace_all(text, "conferencewas", "conference was"),
         text = str_replace_all(text, "conferencewith", "conference with"),
         text = str_replace_all(text, "influenceof", "influence of"),
         text = str_replace_all(text, "evidencethat", "evidence that"),
         text = str_replace_all(text, "residenceproperty", "residence property"),
         text = str_replace_all(text, "\\bagricul tural", "agricultural"),
         text = str_replace_all(text, "\\bcul tural", "cultural"),
         text = str_replace_all(text, "culnj tural", "cultural"),
         text = str_replace_all(text, "struc tural", "structural"),
         text = str_replace_all(text, "struc ture", "structure"),
         text = str_replace_all(text, "organi zations", "organizations"),
         text = str_replace_all(text, "authorizationsunappropriated",
                                "authorizations unappropriated"),
         text = str_replace_all(text, "societiesgeneral", "societies general"),
         text = str_replace_all(text, "terminationof", "termination of"),
         text = str_replace_all(text, "wasmade", "was made"),
         text = str_replace_all(text, "emergencyaid", "emergency aid"),
         text = str_replace_all(text, "tomeetapproappropriationspledgesand",
                                "to meet appropriations pledges and"),
         text = str_replace_all(text, "followsbalarice", " follows balance"),
         text = str_replace_all(text, "termwas", "term was"),
         text = str_replace_all(text, "foundationprovides", "foundation provides"),
         text = str_replace_all(text, "generalbudget", "general budget"),
         text = str_replace_all(text, "givensince", "given since"),
         text = str_replace_all(text, "increasingproportion", "increasing proportion"),
         text = str_replace_all(text, "itfinnaly", "it finnaly"),
         text = str_replace_all(text, "assumedentire", "assumed entire"),
         text = str_replace_all(text, "termwas", "term was"),
         text = str_replace_all(text, "ofyellow", "of yellow"),
         text = str_replace_all(text, "studiesin", "studies in"),
         text = str_replace_all(text, "workthe", "work the"),
         text = str_replace_all(text, "programof", "program of"),
         text = str_replace_all(text, "effectivefor", "effective for"),
         text = str_replace_all(text, "intracontmental", "intracontinental"),
         text = str_replace_all(text, "intelli gence", "intelligence"),
         text = str_replace_all(text, "internalmedicine", "internal medicine"),
         text = str_replace_all(text, "forastudyoftheuseof", "for a study of the
                                use of"),
         text = str_replace_all(text, "foundationinternational", "foundation international"),
         text = str_replace_all(text, "mankindthroughout", "mankind throughout"),
         text = str_replace_all(text, "havedemonstrated", "have demonstrated"),
         text = str_replace_all(text, "beendirected", "been directed"),
         text = str_replace_all(text, "todayand", "today and"),
         text = str_replace_all(text, "resiliencein", "resilience in"),
         text = str_replace_all(text, "environmentalproblems", "environmental problems"),
         text = str_replace_all(text, "foodgrainproduction", "food grain production"),
         text = str_replace_all(text, "malariaparasites", "malaria parasites"),
         text = str_replace_all(text, "activitieswhich", "activities which"),
         text = str_replace_all(text, "largescale", "large scale"),
         text = str_replace_all(text, "theinstitute", "the institute"),
         text = str_replace_all(text, "biologyand", "biology and"),
         text = str_replace_all(text, "taxexempt", "tax exempt"),
         text = str_replace_all(text, "economiesand", "economies and"),
         text = str_replace_all(text, "economicshocks", "economic shocks"),
         text = str_replace_all(text, "accompanyingstatements", "accompanying statements"),
         text = str_replace_all(text, "payableduring", "payable during"),
         text = str_replace_all(text, "agriculturalrsearch", "agricultural research"),
         text = str_replace_all(text, "trainingprogram", "training program"),
         text = str_replace_all(text, "foundationhave", "foundation have"),
         text = str_replace_all(text, "twoyear", "two year"),
         text = str_replace_all(text, "councilcommittee", "council committee"),
         text = str_replace_all(text, "afinalappropriation", "a final appropriation"),
         text = str_replace_all(text, "institutebiochemistry", "institute biochemistry"),
         text = str_replace_all(text, "threeyearperiod", "three year period"),
         text = str_replace_all(text, "completefinancialstatement", "complete financial statement"),
         text = str_replace_all(text, "foundationrepresents", "foundation represents"),
         text = str_replace_all(text, "tenyearperiod", "ten year period"),
         text = str_replace_all(text, "economichandicaps", "economic handicaps"),
         text = str_replace_all(text, "foreignlanguage", "foreign language"),
         text = str_replace_all(text, "appropriationsfor", "appropriations for"),
         text = str_replace_all(text, "carefullyselected", "carefully selected"),
         text = str_replace_all(text, "chinamedical", "china medical"),
         text = str_replace_all(text, "hokkaidouniversity", "hokkaido university"),
         text = str_replace_all(text, "nationalagricultural", "national agricultural"),
         text = str_replace_all(text, "likelyto", "likely to"),
         text = str_replace_all(text, "socialand", "social and"),
         text = str_replace_all(text, "mobilityregional", "mobility regional"),
         text = str_replace_all(text, "medicalscience", "medical science"),
         text = str_replace_all(text, "beendeveloped", "been developed"),
         text = str_replace_all(text, "includingthe", "including the"),
         text = str_replace_all(text, "andeducation", "and education"),
         text = str_replace_all(text, "advancementof", "advancement of"),
         text = str_replace_all(text, "appropriationsgrants", "appropriations grants"),
         text = str_replace_all(text, "applicantsthe", "applicants the"),
         text = str_replace_all(text, "sectorinnovation", "sector innovation"),
         text = str_replace_all(text, "organizationalbreakthroughs", "organizational breakthroughs"),
         text = str_replace_all(text, "localcommunities", "local communities"),
         text = str_replace_all(text, "coricentrates", "concentrates"),
         text = str_replace_all(text, "cerealresearch", "cereal research"),
         text = str_replace_all(text, "periodadriano", "period adriano"),
         text = str_replace_all(text, "threeyear", "three year"),
         text = str_replace_all(text, "shortterm", "short term"),
         text = str_replace_all(text, "closecooperation", "close cooperation"),
         text = str_replace_all(text, "artistinresidence", "artist in residence"),
         text = str_replace_all(text, "wellbeing(complexities|yet)", "wellbeing \\1"),
         text = str_replace_all(text, "welfare(budgets|in|philip|reform|related|to|work)", "welfare \\1"),
         text = str_replace_all(text, "(community|health|instituteofchild)welfare(recreation|rf)?", "\\1 welfare \\2"),
         text = str_replace_all(text, "instituteofchild", "institute of child"),
         text = str_replace_all(text, "fellowshipsandgrantinatdmpeipingunionmedicalcoilegerf", "fellowships and grant in aid in peiping union medical college"),
         text = str_replace_all(text, "thefieldof", "the field of"),
         text = str_replace_all(text, "paraly sis", "paralysis"),
         text = str_replace_all(text, "tuberculo sis", "tuberculosis"),
         text = str_replace_all(text, "mito sis", "mitosis"),
         text = str_replace_all(text, "morphogene sis", "morphogenesis"),
         text = str_replace_all(text, "empha sis", "emphasis"),
         text = str_replace_all(text, "synthe sis", "synthesis"),
         text = str_replace_all(text, "jniversity", "university"),
         text = str_replace_all(text, "jniversity", "university"),
         text = str_replace_all(text, "rcfkefeller", "rockefeller"),
         text = str_replace_all(text, "hworm", "hookworm"),
         text = str_replace_all(text, "\\bsch\\b", "school"),
         text = str_replace_all(text, "appointedfrom", "appointed from"),
         text = str_replace_all(text, "creativityculture", "creativity culture"),
         text = str_replace_all(text, "\\bcoss", "costs"),
         text = str_replace_all(text, "crosstheme", "cross-theme"),
         text = str_replace_all(text, "esearch", "research"),
         text = str_replace_all(text, "schoolistosomiasis", "schistosomiasis"),
         text = str_replace_all(text, "cerate", "coöperate"),
         text = str_replace_all(text, "sciencebased", "science-based"),
         text = str_replace_all(text, "appointedfrom", "appointed from"),
         text = str_replace_all(text, "annualreport", "annual-report"),
         text = str_replace_all(text, "\\bartis ans", "artisans"),
         text = str_replace_all(text, "\\bartis anal", "artisanal"),
         text = str_replace_all(text, "\\bagron omy", "agronomy"),
         text = str_replace_all(text, "\\bagron\\b", "agronomy"),
         text = str_replace_all(text, "\\brresearch", "research"),
         text = str_replace_all(text, "boredhole", "bored-hole"),
         text = str_replace_all(text, "collegeof", "college of"),
         text = str_replace_all(text, "cciences", "sciences"),
         text = str_replace_all(text, "intellectualproperty", "intellectual-property"),
         text = str_replace_all(text, "technologytransfer", "technology transfer"),
         text = str_replace_all(text, "sciencebased", "science-based"),
         text = str_replace_all(text, "\\banal ysis", "analysis"),
         text = str_replace_all(text, "\\bana \\| ysis", "analysis"),
         text = str_replace_all(text, "\\bana ysis", "analysis"),
         text = str_replace_all(text, "\\bana lyze", "analyze"),
         text = str_replace_all(text, "centra", "centro"),
         text = str_replace_all(text, "childh", "childhood"),
         text = str_replace_all(text, "hong kong", "hong-kong"),
         text = str_replace_all(text, "\\bana lyze", "analyze"),
         text = str_replace_all(text, "prob lem(s?)", "problem"),
         text = str_replace_all(text, "prob ably", "probably"),
         text = str_replace_all(text, "cerating", "cooperating"),
         text = str_replace_all(text, "concentrotion", "concentration"),
         text = str_replace_all(text, "\\bnternational", "international"),
         text = str_replace_all(text, "\\bnternationa", "international"),
         text = str_replace_all(text, "middleincome", "middle-income"),
         text = str_replace_all(text, "coriceptual", "conceptual"),
         text = str_replace_all(text, "fundingfor", "funding for"),
         text = str_replace_all(text, "partlclpatlon", "participation"),
         text = str_replace_all(text, "actlvlties", "activities"),
         text = str_replace_all(text, "instttute", "institute"),
         text = str_replace_all(text, "lts", "its"),
         text = str_replace_all(text, "bepobt", "report"),
         text = str_replace_all(text, "foundationnew", "foundation new"),
         text = str_replace_all(text, "carreldakin", "carrel-dakin"),
         text = str_replace_all(text, "cerative", "cooperative"),
         text = str_replace_all(text, "artistinresidence", "artist in residence"),
         text = str_replace_all(text, "andfor", "and for"),
         text = str_replace_all(text, "trainingfor", "training for"),
         text = str_replace_all(text, "distributionand", "distribution and"),
         text = str_replace_all(text, "communitybased", "community based"),
         text = str_replace_all(text, "programrelated", "program related"),
         text = str_replace_all(text, "climateconscious", "climate-conscious"),
         text = str_replace_all(text, "climaterelated", "climate-related"),
         text = str_replace_all(text, "nongovernmental", "non-governmental"),
         text = str_replace_all(text, "educatlon", "education"),
         text = str_replace_all(text, "materlals", "non-governmental"),
         text = str_replace_all(text, "chlldcare", "childcare"),
         text = str_replace_all(text, "afncan", "african"),
         text = str_replace_all(text, "socialservice", "social-service"),
         text = str_replace_all(text, "nonprofit", "non-profit"),
         text = str_replace_all(text, "theformation", "the formation"),
         text = str_replace_all(text, "towardthe", "toward the"),
         text = str_replace_all(text, "fiscalequitynew", "fiscal equity new"),
         text = str_replace_all(text, "annouriced", "announced"),
         text = str_replace_all(text, "\\bmem\\sber(s)?", "member\\1"),
         text = str_replace_all(text, "resuits", "results"),
         text = str_replace_all(text, "centrol", "central"),
         text = str_replace_all(text, "ofthe", "of the"))

beep("ready")

##Cleaning and creating the data with punctuation ----
data.clean.punct <- Corpus  %>%
  mutate(text = str_replace_all(text, "\\n", " "),
         # text = str_remove_all(text, "[:punct:]"), # this line is doing the exact same as the 33, right?
         text = str_remove_all(text, "[:digit:]"), # removes digits
         text = str_remove_all(text, "_+"), # removes underscores
         text = str_remove_all(text, "(\\$|\\~|c~|\\^|\\£|\\^|<>|<|>|¯|>>|⇒|→|>oo|\\boo\\b|°|±|™|\\+|\\-|\\¥|\\■|\\|)\\w?"), # removes specific symbols and patterns
         text = tolower(text), # converts uppercases to lowercases
         text = str_replace_all(text, "wellbeing", "well-being"),
         text = str_remove_all(text,"\\bwe the rockefeller foundation\\b"),
         text = str_remove_all(text,"wwwrockfoundorg"),
         text = str_remove_all(text,"wwwrockefeller"),
         text = str_replace_all(text, "erice(s)?\\b", "ence\\1"),
         text = str_replace_all(text, "univer\\s*(.|usa)?\\s*sity", "university"),
         text = str_replace_all(text, "\\b(organi|moderni)\\s+(.|usa)?\\s+zation", "\\1zation"),
         text = str_replace_all(text, "appropria(\\s|\\t|\\n)+tions", "appropriations"),
         text = str_replace_all(text, "\\b(ac|ag|aq)\\spriations", "\\1priations"),
         text = str_remove_all(text, "\\brockefeller foundation\\b"),
         text = str_replace_all(text, "\\b(adjust|equip)\\s\\b[pjo]\\b", "\\1"),
         text = str_replace_all(text, "nonallgn", "non alligned"),
         text = str_replace_all(text, "(govern|depart|treat|equip|develop|require|move|impove|endow|employ|install|agree|
         establish|disburse|adjust|supple|experi|settle|pay|invest|ele|achieve|commit|appoint|docu|arrange|assign)\\sment", "\\1ment"),
         text = str_replace_all(text, "\\b([a-z]{1,12})\\s\\bment(s?)\\b", "\\1ment\\2"),
         text = str_replace_all(text, "devel\\sopment", "development"),
         text = str_replace_all(text, "de\\svelopment", "development"),
         text = str_replace_all(text, " depart\\sments", "departments"),
         text = str_replace_all(text, "\\b([a-z]{0,9})\\s\\bcation(s?)\\b", "\\1cation\\2"),
         text = str_replace_all(text, "telecommun", "telecommuni"),
         text = str_replace_all(text, "\\b([a-z]{1,10})\\slation(s?)", "\\1lation\\2"),
         text = str_replace_all(text, "foun\\sdation", "foundation"),
         text = str_replace_all(text, "accomo\\sdation", "accomodation"),
         text = str_replace_all(text, "comis\\sj", "comis"),
         text = str_replace_all(text, "([a-z]{2,9})\\ssion(s?)", "\\1ssion\\2"),
         text = str_replace_all(text, "([a-z]{2,12})\\sion(s?)", "\\1sion\\2"),
         text = str_replace_all(text, "foun\\s(a|i)", "foun"),
         text = str_replace_all(text, "con\\struction", "construction"),
         text = str_replace_all(text, "oo struction", "construction"),
         text = str_replace_all(text, "indebted\\s(a|k)", "indebted"),
         text = str_replace_all(text, "([a-z]{0,11})\\s\\bness(e?s?)\\b", "\\1ness\\2"),
         text = str_replace_all(text, "feebleminded ness", "feeblemindedness"),
         text = str_replace_all(text, "\\bcom\\sq", "com"),
         text = str_replace_all(text, "\\bcom\\smunit(y|ies)", "communit\\1"),
         text = str_replace_all(text, "im\\smunity", "immunity"),
         text = str_replace_all(text, "corpo\\s~", "corpo"),
         text = str_replace_all(text, "\\b([a-z]{0,9})\\s\\bration(s?)\\b", "\\1ration\\2"),
         text = str_replace_all(text, "\\badminis\\s>\\sj", "adminis"),
         text = str_replace_all(text, "\\b([a-z]{2,9})\\s\\btration(s?)\\b", "\\1tration\\2"),
         text = str_replace_all(text, "([a-z]{2,10})\\s\\btional(s?)\\b", "\\1tional\\2"),
         text = str_replace_all(text, "\\b([a-z]{2,10})\\s\\btural(s?)\\b", "\\1tural\\2"),
         text = str_replace_all(text, "\\b([a-z]{1,11})\\s\\bation(s?)\\b", "\\1ation\\2"),
         text = str_replace_all(text, "\\b([a-z]{1,10})\\s\\btation(s?)\\b", "\\1tation\\2"),
         text = str_replace_all(text, "\\b([a-z]{2,12})\\s\\btion(s?)\\b", "\\1tion\\2"),
         text = str_replace_all(text, "\\b([a-z]{1,13})\\s\\bit(y|ies)\\b", "\\1it\\2"),
         text = str_replace_all(text, "([a-z]{1,12})\\s\\bence(s?)\\b", "\\1ence\\2"),
         text = str_replace_all(text, "\\b([a-z]{2,11})\\s\\btive(s?)\\b", "\\1tive\\2"),
         text = str_replace_all(text, "attenu ated", "attenuated"),
         text = str_replace_all(text, "\\boper ated", "operated"),
         text = str_replace_all(text, "appropri ated", "appropriated"),
         text = str_replace_all(text, "initi ated", "initiated"),
         text = str_replace_all(text, "evalu ated", "evaluated"),
         text = str_replace_all(text, "cre ated", "created"),
         text = str_replace_all(text, "associ ated", "associated"),
         text = str_replace_all(text, "situ ated", "situated"),
         text = str_replace_all(text, "affili oo ated", "affiliated"),
         text = str_replace_all(text, "cooper ated", "cooperated"),
         text = str_replace_all(text, "unaffili", "unaffiliated"),
         text = str_replace_all(text, "exagger ated", "exaggerated"),
         text = str_replace_all(text, "jniver ~", "univers"),
         text = str_replace_all(text, "ca lamities", "calamities"),
         text = str_replace_all(text, "accaountantfacil", "accaountant facil "),
         text = str_replace_all(text, "popu atlon", "population"),
         text = str_replace_all(text, "expendi tures", "expenditures"),
         text = str_replace_all(text, "corpoi", "corpor"),
         text = str_replace_all(text, "oper >", "oper"),
         text = str_replace_all(text, "collaboi", "collabor"),
         text = str_replace_all(text, "pos sible", "possible"),
         text = str_replace_all(text, "ap propri", "appropri"),
         text = str_replace_all(text, "populationappropri", "population appropri"),
         text = str_replace_all(text, "admin istration", "administration"),
         text = str_replace_all(text, "expendi ", "expenditures"),
         text = str_replace_all(text, "cd room", "cd-room"),
         text = str_replace_all(text, "nce", "rice"),
         text = str_replace_all(text, "viricent", "vincent"),
         text = str_replace_all(text, "organizuion", "organization"),
         text = str_replace_all(text, "autnoma", "autonoma"),
         text = str_replace_all(text, "ommunities", "communities"),
         text = str_replace_all(text, "ommunity", "community"),
         text = str_replace_all(text, "insurarice", "insurance"),
         text = str_replace_all(text, "ommunities", "communities"),
         text = str_replace_all(text, "californiaa", "california"),
         text = str_replace_all(text, "cccommunity", "community"),
         text = str_replace_all(text, "ccommunity", "community"),
         text = str_replace_all(text, "cccommunities", "communities"),
         text = str_replace_all(text, "ccommunities", "communities"),
         text = str_replace_all(text, "certifi cates", "certificates"),
         text = str_replace_all(text, "cer tificates", "certificates"),
         text = str_replace_all(text, "theadvancement", "the advancement"),
         text = str_replace_all(text, "thedevelopment", "the development"),
         text = str_replace_all(text, "thedivision", "the division"),
         text = str_replace_all(text, "thedirection", "the direction"),
         text = str_replace_all(text, "thegovernment", "the government"),
         text = str_replace_all(text, "thecommission", "the commission"),
         text = str_replace_all(text, "theestablishment", "the establishment"),
         text = str_replace_all(text, "thepopulation", "the population"),
         text = str_replace_all(text, "theimprovement", "the improvement"),
         text = str_replace_all(text, "thereproduction", "the reproduction"),
         text = str_replace_all(text, "theassociation", "the association"),
         text = str_replace_all(text, "thetreatment", "the treatment"),
         text = str_replace_all(text, "thepreparation", "the preparation"),
         text = str_replace_all(text, "theadvancement", "the advancement"),
         text = str_replace_all(text, "theorganization", "the organization"),
         text = str_replace_all(text, "thecorporation", "the corporation"),
         text = str_replace_all(text, "thedistribution", "the distribution"),
         text = str_replace_all(text, "theintroduction", "the introduction"),
         text = str_replace_all(text, "thecreation", "the creation"),
         text = str_replace_all(text, "thecooperation", "the cooperation"),
         text = str_replace_all(text, "theinfection", "the infection"),
         text = str_replace_all(text, "theconstruction", "the construction"),
         text = str_replace_all(text, "theapplication", "the application"),
         text = str_replace_all(text, "thesolution", "the solution"),
         text = str_replace_all(text, "theinvestigation", "the investigation"),
         text = str_replace_all(text, "thecompletion", "the completion"),
         text = str_replace_all(text, "thestation", "the station"),
         text = str_replace_all(text, "thepromotion", "the promotion"),
         text = str_replace_all(text, "theuniversity", "the university"),
         text = str_replace_all(text, "thefoundation", "the foundation"),
         text = str_replace_all(text, "thedepartment", "the department"),
         text = str_replace_all(text, "thefoundations", "the foundations"),
         text = str_replace_all(text, "thehumanities", "the humanities"),
         text = str_replace_all(text, "thecity", "the city"),
         text = str_replace_all(text, "theappropriations", "the appropriations"),
         text = str_replace_all(text, "therockefeller", "the rockefeller"),
         text = str_replace_all(text, "theeditors", "the editors"),
         text = str_replace_all(text, "thegrants", "the grants"),
         text = str_replace_all(text, "theblood", "the blood"),
         text = str_replace_all(text, "thecommittee", "the committee"),
         text = str_replace_all(text, "thebureau", "the bureau"),
         text = str_replace_all(text, "theimportant", "the important"),
         text = str_replace_all(text, "theressources", "the ressources"),
         text = str_replace_all(text, "witheressources", "the ressources"),
         text = str_replace_all(text, "theinternational", "the international"),
         text = str_replace_all(text, "thepractice", "the practice"),
         text = str_replace_all(text, "theproject", "the project"),
         text = str_replace_all(text, "theworld", "the world"),
         text = str_replace_all(text, "theinternal", "the internal"),
         text = str_replace_all(text, "thecouncil", "the council"),
         text = str_replace_all(text, "theeconomic", "the economic"),
         text = str_replace_all(text, "thenational", "the national"),
         text = str_replace_all(text, "themedical", "the medical"),
         text = str_replace_all(text, "thework", "the work"),
         text = str_replace_all(text, "thesocial", "the social"),
         text = str_replace_all(text, "thecalifornia", "the california"),
         text = str_replace_all(text, "thestate", "the state"),
         text = str_replace_all(text, "theothers", "the others"),
         text = str_replace_all(text, "andequipment", "and equipment"),
         text = str_replace_all(text, "andtreatment", "and treatment"),
         text = str_replace_all(text, "andgovernment", "and government"),
         text = str_replace_all(text, "anddistribution", "and distribution"),
         text = str_replace_all(text, "andfixedequipment", "and fixed equipement"),
         text = str_replace_all(text, "anddevelopment", "and development"),
         text = str_replace_all(text, "andalso", "and also"),
         text = str_replace_all(text, "andcommodities", "and commodities"),
         text = str_replace_all(text, "andeconomic", "and economic"),
         text = str_replace_all(text, "andwomen", "and women"),
         text = str_replace_all(text, "andrelationships", "and relationships"),
         text = str_replace_all(text, "andscientists", "and scientists"),
         text = str_replace_all(text, "andpossibilities", "and possibilities"),
         text = str_replace_all(text, "andexpanded", "and expanded"),
         text = str_replace_all(text, "andaccordingly", "and accordingly"),
         text = str_replace_all(text, "ofinfection", "of infection"),
         text = str_replace_all(text, "ofdevelopment", "of development"),
         text = str_replace_all(text, "offoundation", "of foundation"),
         text = str_replace_all(text, "ofpopulation", "of population"),
         text = str_replace_all(text, "ofequipment", "of equipment"),
         text = str_replace_all(text, "oftreatment", "of treatment"),
         text = str_replace_all(text, "ofnutrition", "of nutrition"),
         text = str_replace_all(text, "ofinvestigations", "of investigations"),
         text = str_replace_all(text, "ofinformation", "of informtion"),
         text = str_replace_all(text, "ofrecognized", "of recognised"),
         text = str_replace_all(text, "ofmedical", "of medical"),
         text = str_replace_all(text, "ofeducation", "of education"),
         text = str_replace_all(text, "ofminorityyouth", "of minority youth"),
         text = str_replace_all(text, "ofhealth", "of health"),
         text = str_replace_all(text, "ofdistinguished", "of distinguished"),
         text = str_replace_all(text, "ofsanitary", "of sanitary"),
         text = str_replace_all(text, "ofresearch", "of research"),
         text = str_replace_all(text, "overthe", "over the"),
         text = str_replace_all(text, "withtheressources", "with the ressources"),
         text = str_replace_all(text, "withmore", "with more"),
         text = str_replace_all(text, "withfoundation", "with foundation"),
         text = str_replace_all(text, "toaccomplish", "to accomplish"),
         text = str_replace_all(text, "tooperate", "to operate"),
         text = str_replace_all(text, "toinequality", "to inequality"),
         text = str_replace_all(text, "toincrease", "to increase"),
         text = str_replace_all(text, "tothe", "to the"),
         text = str_replace_all(text, "toassist", "to assist"),
         text = str_replace_all(text, "aneducation", "an education"),
         text = str_replace_all(text, "anappropriation", "an appropriation"),
         text = str_replace_all(text, "theparticipating", "the participating"),
         text = str_replace_all(text, "afoundation", "a foundation"),
         text = str_replace_all(text, "asingle", "a single"),
         text = str_replace_all(text, "whetherthey", "wether they"),
         text = str_replace_all(text, "foodproduction", "food production"),
         text = str_replace_all(text, "incollaboration", "in collaboration"),
         text = str_replace_all(text, "incooperation", "in cooperation"),
         text = str_replace_all(text, "inaddition", "in addition"),
         text = str_replace_all(text, "infieldof", "in field of"),
         text = str_replace_all(text, "fordevelopment", "for development"),
         text = str_replace_all(text, "forcollaborative", "for collaborative"),
         text = str_replace_all(text, "foruse", "for use"),
         text = str_replace_all(text, "forthe", "for the"),
         text = str_replace_all(text, "healthdivision", "health division"),
         text = str_replace_all(text, "healthorganization" , "health organization"),
         text = str_replace_all(text, "healthdepartment", "health department"),
         text = str_replace_all(text, "healtheducation", "health education"),
         text = str_replace_all(text, "healthadministration", "health administration"),
         text = str_replace_all(text, "healthcommission", "health commission"),
         text = str_replace_all(text, "highereducation", "higher education"),
         text = str_replace_all(text, "medicaleducation", "medical education"),
         text = str_replace_all(text, "publicadministration", "public administration"),
         text = str_replace_all(text, "foundationfoundation", "foundation"),
         text = str_replace_all(text, "foundationtable", "foundation table"),
         text = str_replace_all(text, "foundationcontents", "foundation contents"),
         text = str_replace_all(text, "foundationgrants", "foundation grants"),
         text = str_replace_all(text, "foundationuniversity", "foundation university"),
         text = str_replace_all(text, "foundationscience", "foundation science"),
         text = str_replace_all(text, "foundationintroductionthe", "foundation
                                introduction the"),
         text = str_replace_all(text, "foundationsystematically", "foundation systematically"),
         text = str_replace_all(text, "foundationsupported", "foundation supported"),
         text = str_replace_all(text, "foundationgrant", "foundation grant"),
         text = str_replace_all(text, "foundationabout", "foundation about"),
         text = str_replace_all(text, "foundationin", "foundation in"),
         text = str_replace_all(text, "foundationadministered", "foundation administered"),
         text = str_replace_all(text, "foundationtion", "foundation"),
         text = str_replace_all(text, "foundationphotograph", "foundation photograph"),
         text = str_replace_all(text, "foundationon", "foundation on"),
         text = str_replace_all(text, "foundationbalarice", "foundation balance"),
         text = str_replace_all(text, "foundationopportunities", "foundation opportunities"),
         text = str_replace_all(text, "soilpollution", "soil pollution"),
         text = str_replace_all(text, "soilshokkaido", "soils hokkaido"),
         text = str_replace_all(text, "wuniversity", "university"),
         text = str_replace_all(text, "universitythe", "university the"),
         text = str_replace_all(text, "univesitythe", "university the"),
         text = str_replace_all(text, "umversidad", "universidad"),
         text = str_replace_all(text, "communitydevelopment", "community development"),
         text = str_replace_all(text, "hookworminfection", "hookworm infection"),
         text = str_replace_all(text, "economicdevelopment", "economic development"),
         text = str_replace_all(text, "reportfoundation", "report foundation"),
         text = str_replace_all(text, "nursingeducation", "nursing education"),
         text = str_replace_all(text, "educationagriculture", "education agriculture"),
         text = str_replace_all(text, "specialattention", "special attention"),
         text = str_replace_all(text, "sciencenutrition", "science nutrition"),
         text = str_replace_all(text, "ruralreconstruction", "rural reconstruction"),
         text = str_replace_all(text, "complementfixation", "complement fixation"),
         text = str_replace_all(text, "saltflotation", "salt flotation"),
         text = str_replace_all(text, "rockfellerfoundation", "rockfeller foundation"),
         text = str_replace_all(text, "perioduniversity", "period university"),
         text = str_replace_all(text, "grantsuniversity", "grants university"),
         text = str_replace_all(text, "internationalrelations", "international relations"),
         text = str_replace_all(text, "guniversity", "university"),
         text = str_replace_all(text, "columbiauniversity", "columbia university"),
         text = str_replace_all(text, "hssuniversity", "hss university"),
         text = str_replace_all(text, "paymentsuniversities", "payments universities"),
         text = str_replace_all(text, "forfiveyears", "for five years"),
         text = str_replace_all(text, "highyielding", "high yielding"),
         text = str_replace_all(text, "highquality", "high quality"),
         text = str_replace_all(text, "arthropodborne", "arthropod-borne"),
         text = str_replace_all(text, "nee", "rice"),
         text = str_replace_all(text, "hivaids", "hiv aids"),
         text = str_replace_all(text, "tne", "the"),
         text = str_replace_all(text, "latrines", "latrine"),
         text = str_replace_all(text, "otherexhibition", "other exhibition"),
         text = str_replace_all(text, "rockefellerfoundation", "rockefeller foundation"),
         text = str_replace_all(text, "harvarduniversity", "harvard university"),
         text = str_replace_all(text, "aboveuniversity", "above university"),
         text = str_replace_all(text, "generaleducation", "general education"),
         text = str_replace_all(text, "nonproliferation", "non-proliferation"),
         text = str_replace_all(text, "periodunivkhsity", "period university"),
         text = str_replace_all(text, "varietiestanzania", "varieties tanzania"),
         text = str_replace_all(text, "policieshuman", "policies human"),
         text = str_replace_all(text, "salaamtanzania", "salaam tanzania"),
         text = str_replace_all(text, "salaamfrom", "salaam from"),
         text = str_replace_all(text, "ascomprehensive", "as comprehensive"),
         text = str_replace_all(text, "characterizationbiotechnologyscnpps",
                                "characterization biotechnology scnpps"),
         text = str_replace_all(text, "associatepatricia", "associate patricia"),
         text = str_replace_all(text, "associatejohn", "associate john"),
         text = str_replace_all(text, "biologyjudge", "biology judge"),
         text = str_replace_all(text, "minoritygroup", "minority group"),
         text = str_replace_all(text, "europeanracial", "european racial"),
         text = str_replace_all(text, "mexicanamerican", "mexican american"),
         text = str_replace_all(text, "peruuniversity", "peru university"),
         text = str_replace_all(text, "japanhokkaido", "japan hokkaido"),
         text = str_replace_all(text, "japangraduate", "japan graduate"),
         text = str_replace_all(text, "japanhague", "japan hague"),
         text = str_replace_all(text, "japaneseunited", "japanese united"),
         text = str_replace_all(text, "colombiauniversity", "colombia university"),
         text = str_replace_all(text, "colombiaversity", "colombia university"),
         text = str_replace_all(text, "brazillaboratory", "brazil laboratory"),
         text = str_replace_all(text, "availableto", "available to"),
         text = str_replace_all(text, "fourfellowships", "four fellowships"),
         text = str_replace_all(text, "valuespririceton", "values princenton"),
         text = str_replace_all(text, "medicineuniversity", "medicine university"),
         text = str_replace_all(text, "facultyof", "faculty of"),
         text = str_replace_all(text, "veterinarymedicine", "veterinary medicine"),
         text = str_replace_all(text, "biologygenetics", "biologoy genetics"),
         text = str_replace_all(text, "developmentsustainable", "development sustainable"),
         text = str_replace_all(text, "projectsupport", "project support"),
         text = str_replace_all(text, "autnoma", "autonoma"),
         text = str_replace_all(text, "darice", "dance"),
         text = str_replace_all(text, "entrarice", "entrance"),
         text = str_replace_all(text, "substarices", "substances"),
         text = str_replace_all(text, "importarice", "importance"),
         text = str_replace_all(text, "coricentrating", "concentrating"),
         text = str_replace_all(text, "periodjapanhokkaido", "period japan hokkaido"),
         text = str_replace_all(text, "sociologydemography", "sociology demography"),
         text = str_replace_all(text, "laboratoryteaching", "laboratory teaching"),
         text = str_replace_all(text, "httpwwwhindustantimescomnewsfeedindiaindiacheersasmonsoonarriveshopesofbetterfarmoutputraisedarticleaspx",
                                "http www hindustan times com news feed india india cheersas monsoon arrives hopes of better farm out put raised article aspx"),
         text = str_replace_all(text, "resiliericestrategies", "resilience strategies"),
         text = str_replace_all(text, "broadlyshared", "broadly shared"),
         text = str_replace_all(text, "changewhile", "change while"),
         text = str_replace_all(text, "insurariceproducts", "insurance products"),
         text = str_replace_all(text, "themworldand", "them world an"),
         text = str_replace_all(text, "smallholder", "small holder"),
         text = str_replace_all(text, "climatechange", "climate change"),
         text = str_replace_all(text, "activitiesexhibition", "activities exhibition"),
         text = str_replace_all(text, "testimonyvernacular", "testimony vernacular"),
         text = str_replace_all(text, "collaborationand", "collaboration and"),
         text = str_replace_all(text, "providechroreographic", "provide choreographic"),
         text = str_replace_all(text, "anyonesexpextation", "anyones expectation"),
         text = str_replace_all(text, "specialscholarships", "special scholarships"),
         text = str_replace_all(text, "scholarshipassistarice", "scholarship assistance"),
         text = str_replace_all(text, "automaticdrips", "automatic drips"),
         text = str_replace_all(text, "studiesand", "studies and"),
         text = str_replace_all(text, "trainingbases", "training bases"),
         text = str_replace_all(text, "bloodforming", "blood forming"),
         text = str_replace_all(text, "basicbiological", "basic biological"),
         text = str_replace_all(text, "pathologicalthrough", "pathological through"),
         text = str_replace_all(text, "substariceswhich", "substances which"),
         text = str_replace_all(text, "arrangementshave", "arrangement shave"),
         text = str_replace_all(text, "statesintensive", "states intensive"),
         text = str_replace_all(text, "whichhave", "which have"),
         text = str_replace_all(text, "southernstates", "southern states"),
         text = str_replace_all(text, "morethat", "more that"),
         text = str_replace_all(text, "theirresponsibility", "their responsibility"),
         text = str_replace_all(text, "degreeof", "degree of"),
         text = str_replace_all(text, "entirerepublic", "entire republic"),
         text = str_replace_all(text, "effectivemethods", "effective methods"),
         text = str_replace_all(text, "averagedegree", "average degree"),
         text = str_replace_all(text, "governmentprivate", "government private"),
         text = str_replace_all(text, "betterqualified", "better qualified"),
         text = str_replace_all(text, "residualspraying", "residual spraying"),
         text = str_replace_all(text, "tworepresentative", "two representative"),
         text = str_replace_all(text, "worldwideusefulness", "worldwide usefullness"),
         text = str_replace_all(text, "favorabletime", "favorable time"),
         text = str_replace_all(text, "methodthe", "method the"),
         text = str_replace_all(text, "authoritieshowever", "authorities however"),
         text = str_replace_all(text, "decreasein", "decrease in"),
         text = str_replace_all(text, "morelimited", "more limited"),
         text = str_replace_all(text, "whichhas", "which has"),
         text = str_replace_all(text, "isan", "is an"),
         text = str_replace_all(text, "truethat", "true that"),
         text = str_replace_all(text, "decadebecomes", "decade becomes"),
         text = str_replace_all(text, "socialsanitary", "social sanitary"),
         text = str_replace_all(text, "theirdistribution", "their distribution"),
         text = str_replace_all(text, "housingfamily", "housing family"),
         text = str_replace_all(text, "problemswhich", "problems which"),
         text = str_replace_all(text, "byfor", "by for"),
         text = str_replace_all(text, "feasabilitypower", "feasability power"),
         text = str_replace_all(text, "prospectthought", "prospect thought"),
         text = str_replace_all(text, "inprogress", "in progress"),
         text = str_replace_all(text, "selfsufficiency", "self sufficiency"),
         text = str_replace_all(text, "providessupport", "provides support"),
         text = str_replace_all(text, "agegroups", "age groups"),
         text = str_replace_all(text, "proposedpersonal", "proposed personal"),
         text = str_replace_all(text, "scholarsphilanthropic", "scholars philanthropic"),
         text = str_replace_all(text, "noeconomic", "no economic"),
         text = str_replace_all(text, "journalismeducationberkeley",
                                "journalism education berkeley"),
         text = str_replace_all(text, "towardsits", "towards its"),
         text = str_replace_all(text, "numberof", "number of"),
         text = str_replace_all(text, "newspaperindustry", "newspaper industry economic"),
         text = str_replace_all(text, "robinsonfoundationnew",
                                "robinson foundation new"),
         text = str_replace_all(text, "fundadevelopment", "fund a development"),
         text = str_replace_all(text, "citizensexpanding", "citizens expanding"),
         text = str_replace_all(text, "businesscollaboratweprogramsaimedat",
                                "business collaborat we programs aimed at"),
         text = str_replace_all(text, "improvingyouth", "improving youth"),
         text = str_replace_all(text, "eemployment", "employment"),
         text = str_replace_all(text, "testedin", "tested in"),
         text = str_replace_all(text, "careeropportunities", "career opportunities"),
         text = str_replace_all(text, "technologyand", "technology and"),
         text = str_replace_all(text, "developmentsaffecting", "developments affecting"),
         text = str_replace_all(text, "givingsupport", "giving support"),
         text = str_replace_all(text, "sciencesmathematics", "sciences mathematics"),
         text = str_replace_all(text, "assistarice", "assistance"),
         text = str_replace_all(text, "iriception", "inception"),
         text = str_replace_all(text, "coricentrated", "concentrated"),
         text = str_replace_all(text, "coriception", "conception"),
         text = str_replace_all(text, "coricerned", "concerned"),
         text = str_replace_all(text, "sirice", "since"),
         text = str_replace_all(text, "coricentration", "concentration"),
         text = str_replace_all(text, "socialadvarice", "social advance"),
         text = str_replace_all(text, "comparativecommunity", "comparative community"),
         text = str_replace_all(text, "ministration", "administration"),
         text = str_replace_all(text, "appf priation", "appropriation"),
         text = str_replace_all(text, "appro priation", "appropriation"),
         text = str_replace_all(text, "priation", "appropriation"),
         text = str_replace_all(text, "internationalhealth", "international health"),
         text = str_replace_all(text, "publichealth", "public health"),
         text = str_replace_all(text, "interna tional", "international"),
         text = str_replace_all(text, "addi tional", "additional"),
         text = str_replace_all(text, "aroundadvarices", "around advances"),
         text = str_replace_all(text, "conferencewas", "conference was"),
         text = str_replace_all(text, "conferencewith", "conference with"),
         text = str_replace_all(text, "influenceof", "influence of"),
         text = str_replace_all(text, "evidencethat", "evidence that"),
         text = str_replace_all(text, "residenceproperty", "residence property"),
         text = str_replace_all(text, "\\bagricul tural", "agricultural"),
         text = str_replace_all(text, "\\bcul tural", "cultural"),
         text = str_replace_all(text, "culnj tural", "cultural"),
         text = str_replace_all(text, "struc tural", "structural"),
         text = str_replace_all(text, "struc ture", "structure"),
         text = str_replace_all(text, "organi zations", "organizations"),
         text = str_replace_all(text, "authorizationsunappropriated",
                                "authorizations unappropriated"),
         text = str_replace_all(text, "societiesgeneral", "societies general"),
         text = str_replace_all(text, "terminationof", "termination of"),
         text = str_replace_all(text, "wasmade", "was made"),
         text = str_replace_all(text, "emergencyaid", "emergency aid"),
         text = str_replace_all(text, "tomeetapproappropriationspledgesand",
                                "to meet appropriations pledges and"),
         text = str_replace_all(text, "followsbalarice", " follows balance"),
         text = str_replace_all(text, "termwas", "term was"),
         text = str_replace_all(text, "foundationprovides", "foundation provides"),
         text = str_replace_all(text, "generalbudget", "general budget"),
         text = str_replace_all(text, "givensince", "given since"),
         text = str_replace_all(text, "increasingproportion", "increasing proportion"),
         text = str_replace_all(text, "itfinnaly", "it finnaly"),
         text = str_replace_all(text, "assumedentire", "assumed entire"),
         text = str_replace_all(text, "termwas", "term was"),
         text = str_replace_all(text, "ofyellow", "of yellow"),
         text = str_replace_all(text, "studiesin", "studies in"),
         text = str_replace_all(text, "workthe", "work the"),
         text = str_replace_all(text, "programof", "program of"),
         text = str_replace_all(text, "effectivefor", "effective for"),
         text = str_replace_all(text, "intracontmental", "intracontinental"),
         text = str_replace_all(text, "intelli gence", "intelligence"),
         text = str_replace_all(text, "internalmedicine", "internal medicine"),
         text = str_replace_all(text, "forastudyoftheuseof", "for a study of the
                                use of"),
         text = str_replace_all(text, "foundationinternational", "foundation international"),
         text = str_replace_all(text, "mankindthroughout", "mankind throughout"),
         text = str_replace_all(text, "havedemonstrated", "have demonstrated"),
         text = str_replace_all(text, "beendirected", "been directed"),
         text = str_replace_all(text, "todayand", "today and"),
         text = str_replace_all(text, "resiliencein", "resilience in"),
         text = str_replace_all(text, "environmentalproblems", "environmental problems"),
         text = str_replace_all(text, "foodgrainproduction", "food grain production"),
         text = str_replace_all(text, "malariaparasites", "malaria parasites"),
         text = str_replace_all(text, "activitieswhich", "activities which"),
         text = str_replace_all(text, "largescale", "large scale"),
         text = str_replace_all(text, "theinstitute", "the institute"),
         text = str_replace_all(text, "biologyand", "biology and"),
         text = str_replace_all(text, "taxexempt", "tax exempt"),
         text = str_replace_all(text, "economiesand", "economies and"),
         text = str_replace_all(text, "economicshocks", "economic shocks"),
         text = str_replace_all(text, "accompanyingstatements", "accompanying statements"),
         text = str_replace_all(text, "payableduring", "payable during"),
         text = str_replace_all(text, "agriculturalrsearch", "agricultural research"),
         text = str_replace_all(text, "trainingprogram", "training program"),
         text = str_replace_all(text, "foundationhave", "foundation have"),
         text = str_replace_all(text, "twoyear", "two year"),
         text = str_replace_all(text, "councilcommittee", "council committee"),
         text = str_replace_all(text, "afinalappropriation", "a final appropriation"),
         text = str_replace_all(text, "institutebiochemistry", "institute biochemistry"),
         text = str_replace_all(text, "threeyearperiod", "three year period"),
         text = str_replace_all(text, "completefinancialstatement", "complete financial statement"),
         text = str_replace_all(text, "foundationrepresents", "foundation represents"),
         text = str_replace_all(text, "tenyearperiod", "ten year period"),
         text = str_replace_all(text, "economichandicaps", "economic handicaps"),
         text = str_replace_all(text, "foreignlanguage", "foreign language"),
         text = str_replace_all(text, "appropriationsfor", "appropriations for"),
         text = str_replace_all(text, "carefullyselected", "carefully selected"),
         text = str_replace_all(text, "chinamedical", "china medical"),
         text = str_replace_all(text, "hokkaidouniversity", "hokkaido university"),
         text = str_replace_all(text, "nationalagricultural", "national agricultural"),
         text = str_replace_all(text, "likelyto", "likely to"),
         text = str_replace_all(text, "socialand", "social and"),
         text = str_replace_all(text, "mobilityregional", "mobility regional"),
         text = str_replace_all(text, "medicalscience", "medical science"),
         text = str_replace_all(text, "beendeveloped", "been developed"),
         text = str_replace_all(text, "includingthe", "including the"),
         text = str_replace_all(text, "andeducation", "and education"),
         text = str_replace_all(text, "advancementof", "advancement of"),
         text = str_replace_all(text, "appropriationsgrants", "appropriations grants"),
         text = str_replace_all(text, "applicantsthe", "applicants the"),
         text = str_replace_all(text, "sectorinnovation", "sector innovation"),
         text = str_replace_all(text, "organizationalbreakthroughs", "organizational breakthroughs"),
         text = str_replace_all(text, "localcommunities", "local communities"),
         text = str_replace_all(text, "coricentrates", "concentrates"),
         text = str_replace_all(text, "cerealresearch", "cereal research"),
         text = str_replace_all(text, "periodadriano", "period adriano"),
         text = str_replace_all(text, "threeyear", "three year"),
         text = str_replace_all(text, "shortterm", "short term"),
         text = str_replace_all(text, "closecooperation", "close cooperation"),
         text = str_replace_all(text, "artistinresidence", "artist in residence"),
         text = str_replace_all(text, "wellbeing(complexities|yet)", "wellbeing \\1"),
         text = str_replace_all(text, "welfare(budgets|in|philip|reform|related|to|work)", "welfare \\1"),
         text = str_replace_all(text, "(community|health|instituteofchild)welfare(recreation|rf)?", "\\1 welfare \\2"),
         text = str_replace_all(text, "instituteofchild", "institute of child"),
         text = str_replace_all(text, "fellowshipsandgrantinatdmpeipingunionmedicalcoilegerf", "fellowships and grant in aid in peiping union medical college"),
         text = str_replace_all(text, "thefieldof", "the field of"),
         text = str_replace_all(text, "paraly sis", "paralysis"),
         text = str_replace_all(text, "tuberculo sis", "tuberculosis"),
         text = str_replace_all(text, "mito sis", "mitosis"),
         text = str_replace_all(text, "morphogene sis", "morphogenesis"),
         text = str_replace_all(text, "empha sis", "emphasis"),
         text = str_replace_all(text, "synthe sis", "synthesis"),
         text = str_replace_all(text, "jniversity", "university"),
         text = str_replace_all(text, "jniversity", "university"),
         text = str_replace_all(text, "rcfkefeller", "rockefeller"),
         text = str_replace_all(text, "hworm", "hookworm"),
         text = str_replace_all(text, "\\bsch\\b", "school"),
         text = str_replace_all(text, "appointedfrom", "appointed from"),
         text = str_replace_all(text, "creativityculture", "creativity culture"),
         text = str_replace_all(text, "\\bcoss", "costs"),
         text = str_replace_all(text, "crosstheme", "cross-theme"),
         text = str_replace_all(text, "esearch", "research"),
         text = str_replace_all(text, "schoolistosomiasis", "schistosomiasis"),
         text = str_replace_all(text, "cerate", "coöperate"),
         text = str_replace_all(text, "sciencebased", "science-based"),
         text = str_replace_all(text, "appointedfrom", "appointed from"),
         text = str_replace_all(text, "annualreport", "annual-report"),
         text = str_replace_all(text, "\\bartis ans", "artisans"),
         text = str_replace_all(text, "\\bartis anal", "artisanal"),
         text = str_replace_all(text, "\\bagron omy", "agronomy"),
         text = str_replace_all(text, "\\bagron\\b", "agronomy"),
         text = str_replace_all(text, "\\brresearch", "research"),
         text = str_replace_all(text, "boredhole", "bored-hole"),
         text = str_replace_all(text, "collegeof", "college of"),
         text = str_replace_all(text, "cciences", "sciences"),
         text = str_replace_all(text, "intellectualproperty", "intellectual-property"),
         text = str_replace_all(text, "technologytransfer", "technology transfer"),
         text = str_replace_all(text, "sciencebased", "science-based"),
         text = str_replace_all(text, "\\banal ysis", "analysis"),
         text = str_replace_all(text, "\\bana \\| ysis", "analysis"),
         text = str_replace_all(text, "\\bana ysis", "analysis"),
         text = str_replace_all(text, "\\bana lyze", "analyze"),
         text = str_replace_all(text, "centra", "centro"),
         text = str_replace_all(text, "childh", "childhood"),
         text = str_replace_all(text, "hong kong", "hong-kong"),
         text = str_replace_all(text, "\\bana lyze", "analyze"),
         text = str_replace_all(text, "prob lem(s?)", "problem"),
         text = str_replace_all(text, "prob ably", "probably"),
         text = str_replace_all(text, "cerating", "cooperating"),
         text = str_replace_all(text, "concentrotion", "concentration"),
         text = str_replace_all(text, "\\bnternational", "international"),
         text = str_replace_all(text, "\\bnternationa", "international"),
         text = str_replace_all(text, "middleincome", "middle-income"),
         text = str_replace_all(text, "coriceptual", "conceptual"),
         text = str_replace_all(text, "fundingfor", "funding for"),
         text = str_replace_all(text, "partlclpatlon", "participation"),
         text = str_replace_all(text, "actlvlties", "activities"),
         text = str_replace_all(text, "instttute", "institute"),
         text = str_replace_all(text, "lts", "its"),
         text = str_replace_all(text, "bepobt", "report"),
         text = str_replace_all(text, "foundationnew", "foundation new"),
         text = str_replace_all(text, "carreldakin", "carrel-dakin"),
         text = str_replace_all(text, "cerative", "cooperative"),
         text = str_replace_all(text, "artistinresidence", "artist in residence"),
         text = str_replace_all(text, "andfor", "and for"),
         text = str_replace_all(text, "trainingfor", "training for"),
         text = str_replace_all(text, "distributionand", "distribution and"),
         text = str_replace_all(text, "communitybased", "community based"),
         text = str_replace_all(text, "programrelated", "program related"),
         text = str_replace_all(text, "climateconscious", "climate-conscious"),
         text = str_replace_all(text, "climaterelated", "climate-related"),
         text = str_replace_all(text, "nongovernmental", "non-governmental"),
         text = str_replace_all(text, "educatlon", "education"),
         text = str_replace_all(text, "materlals", "non-governmental"),
         text = str_replace_all(text, "chlldcare", "childcare"),
         text = str_replace_all(text, "afncan", "african"),
         text = str_replace_all(text, "socialservice", "social-service"),
         text = str_replace_all(text, "nonprofit", "non-profit"),
         text = str_replace_all(text, "theformation", "the formation"),
         text = str_replace_all(text, "towardthe", "toward the"),
         text = str_replace_all(text, "fiscalequitynew", "fiscal equity new"),
         text = str_replace_all(text, "annouriced", "announced"),
         text = str_replace_all(text, "\\bmem\\sber(s)?", "member\\1"),
         text = str_replace_all(text, "resuits", "results"),
         text = str_replace_all(text, "centrol", "central"),
         text = str_replace_all(text, "ofthe", "of the"))

beep("ready")

## Two ways of testing the objects are different
identical(data.clean.no.punct, data.clean.punct)
all.equal(data.clean.no.punct$text, data.clean.punct$text)


## The best class to see elements within an object is tibble ----
# data.clean.no.punct <- tibble(data.clean.no.punct)
# save(data.clean.no.punct, file = "data.clean.no.punct.Rda")
#
# data.clean.punct <- tibble(data.clean.punct)
# save(data.clean.punct, file = "data.clean.punct.Rda")
