==> 1.1 C.R <==
# 4th Cleaning ----
tokenised.punct <- tokenised.punct %>%
  mutate(words = str_replace_all(words, "approappropriation", "appropriation"),
         words = str_replace_all(words, "approappropriations", "appropriations"),
         words = str_replace_all(words, "approapproriations", "appropriations"),
         words = str_replace_all(words, "adadministration", "administration"),
         words = str_replace_all(words, "engiricer", "engineer"),
         words = str_replace_all(words, "driveroffice", "driver office"),
         words = str_replace_all(words, "workerschallengebetween", "workers challenge between"),
         words = str_replace_all(words, "andworld", "and world"),
         words = str_replace_all(words, "whileeach", "while each"),
         words = str_replace_all(words, "joinother", "join others"),
         words = str_replace_all(words, "successfuldemostrations", "successful demostrations"),
         words = str_replace_all(words, "promoterenewal", "promote renewal"),
         words = str_replace_all(words, "dcnuclear", "dc nuclear"),
         words = str_replace_all(words, "initiativewhashington", "initiative whashington"),
         words = str_replace_all(words, "progressresolved", "progress resolved"),
         words = str_replace_all(words, "campaingfor", "campaingfor"),
         words = str_replace_all(words, "diseasesurveillance", 
                                 "disease surveillance"),
         words = str_replace_all(words, "networkschallengebecause", 
                                 "networks challenge because"),
         words = str_replace_all(words, "resilienceto", "resilience to"),
         words = str_replace_all(words, "forsmall", "for small"),
         words = str_replace_all(words, "impactinvesting", "impact investing"),
         words = str_replace_all(words, "industrychallengegovernment", 
                                 "industry challeng egovernment"),
         words = str_replace_all(words, "stewardshipjohn", "stewardship john"),
         words = str_replace_all(words, "foundationorgcondensed", "foundation org condensed"),
         words = str_replace_all(words, "activitiesdecember", "activities december"),
         words = str_replace_all(words, "rodinpresident", "rodin president"),
         words = str_replace_all(words, "presidentthe", "president the"),
         words = str_replace_all(words, "foundationorggrantsinescapable", 
                                 "foundation org grants inescapable"),
         words = str_replace_all(words, "withmore", "with more"),
         words = str_replace_all(words, "bostonhenry", "boston henry"),
         words = str_replace_all(words, "andthe", "and the"),
         words = str_replace_all(words, "countyhealth", "county health"),
         words = str_replace_all(words, "uponthe", "upon the"),
         words = str_replace_all(words, "wasreopened", "was reopened"),
         words = str_replace_all(words, "boardsrepresentative", 
                                 "boards representative"),
         words = str_replace_all(words, "thesanitary", "the sanitary"),
         words = str_replace_all(words, "weregranted", "were granted"),
         words = str_replace_all(words, "hadbeen", "had been"),
         words = str_replace_all(words, "variousprojects", "various projects"),
         words = str_replace_all(words, "servicesof", "service sof"),
         words = str_replace_all(words, "bythe", "by the"),
         words = str_replace_all(words, "waspermitted", "was permitted"),
         words = str_replace_all(words, "intoconsultation", "into consultation"),
         words = str_replace_all(words, "supplysystem", "supply system"),
         words = str_replace_all(words, "baltimoredonald", "baltimore donald"),
         words = str_replace_all(words, "adviserand", "adviser and"),
         words = str_replace_all(words, "yorkgeorge", "york george"),
         words = str_replace_all(words, "mainte nance", "maintenance"), 
         words = str_replace_all(words, "thevenereal", "the venereal"),
         words = str_replace_all(words, "iscorrelated", "is correlated"),
         words = str_replace_all(words, "epidemiologistwithin", "epidemiologist within"),
         words = str_replace_all(words, "tohonduras", "to honduras"),
         words = str_replace_all(words, "cooperativeprogram", "cooperative program"),
         words = str_replace_all(words, "madesurveys", "made surveys"),
         words = str_replace_all(words, "stationedin", "stationedin"),
         words = str_replace_all(words, "assistancein", "assistance in"),
         words = str_replace_all(words, "operationhookworm", "operation hookworm"),
         words = str_replace_all(words, "atechnical", "a technical"),
         words = str_replace_all(words, "amedical", "amedical"),
         words = str_replace_all(words, "otherplaces", "otherplaces"),
         words = str_replace_all(words, "undergovernmental", "undergovernmental"),
         words = str_replace_all(words, "hehas", "he has"),
         words = str_replace_all(words, "giveadvice", "give advice"),
         words = str_replace_all(words, "thecontrol", "the control"),
         words = str_replace_all(words, "statessanitary", "states sanitary"),
         words = str_replace_all(words, "malariastudies", "malaria studies"),
         words = str_replace_all(words, "servicesof", "services of"),
         words = str_replace_all(words, "ahonduran", "a honduran"),
         words = str_replace_all(words, "ingeneral", "in general"),
         words = str_replace_all(words, "bylending", "by lending"),
         words = str_replace_all(words, "fieldpersonnel", "field personnel"),
         words = str_replace_all(words, "physiciansone", "physicians one"),
         words = str_replace_all(words, "appropriationsa", "appropriations a"),
         words = str_replace_all(words, "healthwith", "health with"),
         words = str_replace_all(words, "underthe", "under the"),
         words = str_replace_all(words, "theservices", "the services"),
         words = str_replace_all(words, "localityin", "localityin"),
         words = str_replace_all(words, "inmany", "in many"),
         words = str_replace_all(words, "availablethe", "available the"),
         words = str_replace_all(words, "drainagesystem", "drainage system"),
         words = str_replace_all(words, "nationaldepartment", "national department"),
         words = str_replace_all(words, "localsanitary", "local sanitary"),
         words = str_replace_all(words, "themthe", "them the"),
         words = str_replace_all(words, "watersupply", "water supply"),
         words = str_replace_all(words, "departmentin", "department in"),
         words = str_replace_all(words, "foundationcontinued", "foundation continued"),
         words = str_replace_all(words, "malariologistand", "malariologist and"),
         words = str_replace_all(words, "theproblem", "the problem"),
         words = str_replace_all(words, "drainagethe", "drainage the"),
         words = str_replace_all(words, "adjacentuntiled", "adjacent untiled"),
         words = str_replace_all(words, "wasoffered", "was offered"),
         words = str_replace_all(words, "foundationssanitary", "foundations sanitary"),
         words = str_replace_all(words, "theunited", "the united"),
         words = str_replace_all(words, "expectedthat", "expected that"),
         words = str_replace_all(words, "hondurasearly", "honduras early"),
         words = str_replace_all(words, "hondurasa", "honduras a"),
         words = str_replace_all(words, "otherplaces", "other places"),
         words = str_replace_all(words, "undergovernmental", "under governmental"),
         words = str_replace_all(words, "servicein", "service in"),
         words = str_replace_all(words, "furnishedby", "furnished by"),
         words = str_replace_all(words, "malariastudies", "malaria studies"),
         words = str_replace_all(words, "controlin", "control in"),
         words = str_replace_all(words, "fordrainage", "for drainage"),
         words = str_replace_all(words, "furthertests", "further tests"),
         words = str_replace_all(words, "divisionin", "division in"),
         words = str_replace_all(words, "nicaraguacontinued", "nicaragua continued"),
         words = str_replace_all(words, "adivision", "a division"),
         words = str_replace_all(words, "governmenthaving", "government having"),
         words = str_replace_all(words, "ofthe ", "of the "),
         words = str_replace_all(words, "representativeserved", "representative served"),
         words = str_replace_all(words, "giveexpert", "give expert"),
         words = str_replace_all(words, "atwoyear", "a two year"),
         words = str_replace_all(words, "andthree", "and three"),
         words = str_replace_all(words, "latelybeen", "lately been"),
         words = str_replace_all(words, "mechanicalequipmentfillinga", 
                                 "mechanical equipment filling a"),
         words = str_replace_all(words, "breedingplace", "breeding place"),
         words = str_replace_all(words, "foundationfor", "foundation for"),
         words = str_replace_all(words, "prepareda", "prepared a"),
         words = str_replace_all(words, "beenlent", "been lent"),
         words = str_replace_all(words, "foundationto", "foundation to"),
         words = str_replace_all(words, "beingtrained", "being trained"),
         words = str_replace_all(words, "foundationfellowship", "foundation fellowship"),
         words = str_replace_all(words, "hondurasand", "honduras and"),
         words = str_replace_all(words, "installinga", "installing a"),
         words = str_replace_all(words, "themthe", "them the"),
         words = str_replace_all(words, "ofceylon", "of ceylon"),
         words = str_replace_all(words, "observationa", "observation a"),
         words = str_replace_all(words, "inthis", "in this"),
         words = str_replace_all(words, "furnishedby", "furnished by"),
         words = str_replace_all(words, "corporationswho", "corporations who"),
         words = str_replace_all(words, "infieldsketchingand", "in field sketching and"),
         words = str_replace_all(words, "sectionof", "section of"),
         words = str_replace_all(words, "maintainsheadquarters", "maintains headquarters"),
         words = str_replace_all(words, "aidedthis", "aided this"),
         words = str_replace_all(words, "asanitarytwoyear", "a sanitary two year"),
         words = str_replace_all(words, "bureauboredhole", "bureau bored hole"),
         words = str_replace_all(words, "departmentsbureauboredhole", 
                                 "departments bureau bored hole"),
         words = str_replace_all(words, "succesfullyinstalled", "succesfully installed"),
         words = str_replace_all(words, "villagesattention", "villages attention"),
         words = str_replace_all(words, "innorth", "in north"),
         words = str_replace_all(words, "practicableroutine", "practicable routine"),
         words = str_replace_all(words, "entomologistcollaborate", "entomologist collaborate"),
         words = str_replace_all(words, "foundationthe", "foundation the"),
         words = str_replace_all(words, "andof", "and of"),
         words = str_replace_all(words, "electroenceph alographyjohns", 
                                 "electroencephalography johns"),   
         words = str_replace_all(words, "europeedinburgh", "europe edinburgh"),
         words = str_replace_all(words, "institutionsgeorgia", "institutions georgia"),
         words = str_replace_all(words, "greecegermany", "greece germany"),
         words = str_replace_all(words, "fultonmontgomery", "fulton montgomery"),
         words = str_replace_all(words, "twonutritionists", "two nutritionists"),
         words = str_replace_all(words, "medicalschools", "medical schools"),
         words = str_replace_all(words, "therequest", "the request"),
         words = str_replace_all(words, "inbarranquilla", "in barranquilla"),
         words = str_replace_all(words, "anentomologist", "an entomologist"),
         words = str_replace_all(words, "staffmember", "staff member"),
         words = str_replace_all(words, "sanitarianthe", "sanitarian the"),
         words = str_replace_all(words, "providingsalaries", "providingsalaries"),
         words = str_replace_all(words, "commissionmade", "commission made"),
         words = str_replace_all(words, "intermittentlythe", "intermittently the"),
         words = str_replace_all(words, "medicalcommission", "medical commission"),
         words = str_replace_all(words, "factorsspecial", "factors special"),
         words = str_replace_all(words, "provedthat", "proved that"),
         words = str_replace_all(words, "thisregion", "this region"),
         words = str_replace_all(words, "wellestablished", "well established"),
         words = str_replace_all(words, "insuresnormal", "insures normal"),
         words = str_replace_all(words, "staffhas", "staff has"),
         words = str_replace_all(words, "antiguamontserrat vincentgrenadines luciadominica", 
                                 "antigua montserrat vincent grenadines lucia dominica"),
         words = str_replace_all(words, "incharge", "in charge"),
         words = str_replace_all(words, "ineurope", "in europe"),
         words = str_replace_all(words, "phaseof", "phase of"),
         words = str_replace_all(words, "researchphysicist", "research physicist"),
         words = str_replace_all(words, "withthe", "with the"),
         words = str_replace_all(words, "agriculturecomprised", "agriculture comprised"),
         words = str_replace_all(words, "anddoctor", "and doctor"),
         words = str_replace_all(words, "andsince", "and since"),
         words = str_replace_all(words, "auxiliaryworkers", "auxiliary workers"),
         words = str_replace_all(words, "subsequentobservation", "subsequent observation"),
         words = str_replace_all(words, "physiciansnurses", "physicians nurses"),
         words = str_replace_all(words, "equipmentgeorge", "equipment george"),
         words = str_replace_all(words, "worksengland", "works england"),
         words = str_replace_all(words, "draftsmen", "draft smen"),
         words = str_replace_all(words, "andsecretaries", "and secretaries"),
         words = str_replace_all(words, "yorkadditional", "york additional"),
         words = str_replace_all(words, "obstetricsand", "obstetrics and"),
         words = str_replace_all(words, "thetheatre", "the theatre"),
         words = str_replace_all(words, "museumlighting", "museum lighting"),
         words = str_replace_all(words, "governmentof", "government of"),
         words = str_replace_all(words, "screeningor", "screening or"),
         words = str_replace_all(words, "formfiltrationhas", "form filtration has"),
         words = str_replace_all(words, "recentbiological", "recent biological"),
         words = str_replace_all(words, "themathematician", "the mathematician"),
         words = str_replace_all(words, "analysisof", "analysis of"),
         words = str_replace_all(words, "lectromagneticfieldwhenan",
                                 "electromagnetic field when an"),
         words = str_replace_all(words, "meanta", "meant a"),
         words = str_replace_all(words, "peopledevelopment", "people development"),
         words = str_replace_all(words, "geneticistarthur", "geneticis tarthur"),
         words = str_replace_all(words, "agriculturewhich", "agriculture which"),
         words = str_replace_all(words, "themathematician", "the mathematician"),
         words = str_replace_all(words, "trainedprofessional", "trained professional"),
         words = str_replace_all(words, "foundationcolombiabogotacolombian", 
                                 "foundation colombia bogota colombian"),
         words = str_replace_all(words, "peopledevelopment", "people development"),
         words = str_replace_all(words, "foundationorgwhat", "foundation what"),
         words = str_replace_all(words, "foundationorgthese", "foundation these"),
         words = str_replace_all(words, "foundationrogduring", "foundation during"),
         words = str_replace_all(words, "theprovirice", "the province"),
         words = str_replace_all(words, "sanitaryengiricering", "sanitary engineering"),
         words = str_replace_all(words, "engiricerprovided", "engineer provided"),
         words = str_replace_all(words, "antimalariaengiricering", "antimalaria engineering"),
         words = str_replace_all(words, "engiriceringdepartment", "engineering department"),
         words = str_replace_all(words, "wellequippedlaboratory", "well equipped laboratory"),
         words = str_replace_all(words, "fundamentalstudies", "fundamental studies"),
         words = str_replace_all(words, "governmentfor", "government for"),
         words = str_replace_all(words, "presidentialdecree", "presidential decree"),
         words = str_replace_all(words, "humangenetics", "human genetics"),
         words = str_replace_all(words, "napleshas thefirstto", "naples has the first to"),
         words = str_replace_all(words, "governmentdevelopment", "government development"),
         words = str_replace_all(words, "environmentalsanitation", "environmental sanitation"),
         words = str_replace_all(words, "connectionwith", "connectio nwith"),
         words = str_replace_all(words, "increasingthe", "increasing the"),
         words = str_replace_all(words, "departmentsof", "departments of"),
         words = str_replace_all(words, "generalhealth", "general health"),
         words = str_replace_all(words, "anddynamic", "and dynamic"),
         words = str_replace_all(words, "asuniversity", "as university"),
         words = str_replace_all(words, "berlindahlem", "berlin dahlem"),
         words = str_replace_all(words, "andlandscape", "and landscape"),
         words = str_replace_all(words, "departmentof", "department of"),
         words = str_replace_all(words, "developmentof", "development of"),
         words = str_replace_all(words, "foundationsciencebased", "foundation science based"),
         words = str_replace_all(words, "atrisk", "at risk"),
         words = str_replace_all(words, "glycerinesalt", "glycerine salt"),
         words = str_replace_all(words, "budgetcontinued", "budget continued"),
         words = str_replace_all(words, "searchlight", "search light"),
         words = str_replace_all(words, "acquisitionloan", "acquisition loan"),
         words = str_replace_all(words, "massachusettsharvard", "massachusetts harvard"),
         words = str_replace_all(words, "medicalboard", "medical board"),
         words = str_replace_all(words, "ofa", "of a"),
         words = str_replace_all(words, "foundationhas", "foundation has"),
         words = str_replace_all(words, "developmgcountry", "developing country"),
         words = str_replace_all(words, "contraceptivesand", "contraceptives and"),
         words = str_replace_all(words, "hereworkinprogress", "here work in progress"),
         words = str_replace_all(words, "kcontinued", "k-continued"),
         words = str_replace_all(words, "paymentsmissionary", "payment smissionary"),
         words = str_replace_all(words, "societieshos pitals", "societies hospitals"),
         words = str_replace_all(words, "ofinternational", "of international"),
         words = str_replace_all(words, "thcentury", "th century"),
         words = str_replace_all(words, "thecountry", "the country"),
         words = str_replace_all(words, "thegeneral", "the general"),
         words = str_replace_all(words, "thhealth", "the health"),
         words = str_replace_all(words, "activitiescontinued", "activities continued"),
         words = str_replace_all(words, "developmentcontinued", "development continued"),
         words = str_replace_all(words, "programscontinued", "programs continued"),
         words = str_replace_all(words, "organizationscontinued", "organizations continued"),
         words = str_replace_all(words, "institutionscontinued", "institutions continued"),
         words = str_replace_all(words, "paymentsgeneral", "payments general"),
         words = str_replace_all(words, "payments research", "payments research"),
         words = str_replace_all(words, "dcontinued", "continued"),
         words = str_replace_all(words, "requirementscontinued", "requirements continued"),
         words = str_replace_all(words, "educationalcontinued", "educational continued"),
         words = str_replace_all(words, "schoolcontinued", "school continued"),
         words = str_replace_all(words, "budgetcontinued", "budget continued"),
         words = str_replace_all(words, "countriescontinued", "countries continued"),
         words = str_replace_all(words, "europecontinued", "europe continued"),
         words = str_replace_all(words, "statescontinued", "states continued"),
         words = str_replace_all(words, "paymentspublic", "payments public"),
         words = str_replace_all(words, "centerfor", "center for"),
         words = str_replace_all(words, "sciencebased", "science based"),
         words = str_replace_all(words, "agncultural", "agricultural"),
         words = str_replace_all(words, "costof", "cost of"),
         words = str_replace_all(words, "costsof", "costs of"),
         words = str_replace_all(words, "amencan", "american"),
         words = str_replace_all(words, "capacitybuilding", "capacity building"),
         words = str_replace_all(words, "civilsocietyorganizations", 
                                 "civil society organizations"),
         words = str_replace_all(words, "fouryear", "four year"),
         words = str_replace_all(words, "droughttolerant", "drought tolerant"),
         words = str_replace_all(words, "appbopria", "appropriations"),
         words = str_replace_all(words, "annourices", "announces"),
         words = str_replace_all(words, "chma", "china"),
         words = str_replace_all(words, "developmentand", "development and"),
         words = str_replace_all(words, "foodand", "food and"),
         words = str_replace_all(words, "erice$", "ence"),
         words = str_replace_all(words, "erices$", "ences"),
         words = str_replace_all(words, "engiricering", "engineering"),
         words = str_replace_all(words, "coricerning", "concerning"),
         words = str_replace_all(words, "(arice)(.*)", "ance\\2"),
         words = str_replace_all(words, "(erice)(.*)", "ence\\2"),
         words = str_replace_all(words, "(irice)(.*)", "ince\\2"),
         words = str_replace_all(words, "(irice)(.*)", "ince\\2"),
         words = str_replace_all(words, "schoololarships", "scholarships"),
         words = str_replace_all(words, "artistinresidence", "artist in residence"),
         words = str_replace_all(words, "byaws", "by-laws"),
         words = str_replace_all(words, "newyorkcity", "new york city"),
         words = str_replace_all(words, "arc", "are"),
         words = str_replace_all(words, "afiveearperiod", "a five year period"),
         words = str_replace_all(words, "beview", "review"),
         words = str_replace_all(words, "composernesidence", "composer-in-residence"),
         words = str_replace_all(words, "inervice", "in-service"),
         words = str_replace_all(words, "\\bdesigna\\b", "designations"),
         words = str_replace_all(words, "berlinahlem", "berlin-dahlem"),
         words = str_replace_all(words, "andexplorations", "and explorations"),
         words = str_replace_all(words, "universityof", "university of"),
         words = str_replace_all(words, "ph.d.", "ph.d"),
         words = str_replace_all(words, "associa", "association"),
         words = str_replace_all(words, "\\bbabwe", "zimbabwe"),
         words = str_replace_all(words, "ogyscnpps", "scripps"),
         words = str_replace_all(words, "biotechnol\\b", "biotechnology"),
         words = str_replace_all(words, "sociol\\b", "sociology"),
         words = str_replace_all(words, "coricert", "concert"),
         words = str_replace_all(words, "ourwork", "our work"),
         words = str_replace_all(words, "arehaeology", "archaeology"),
         words = str_replace_all(words, "appbopbia", "appropriation"),
         words = str_replace_all(words, "researeh", "research"), 
         words = str_replace_all(words, "colomb\\b", "colombia"),
         words = str_replace_all(words, "minorityroup", "minority group"),
         words = str_replace_all(words, "artisthesidence", "artist in residence"),
         words = str_replace_all(words, "avantarde", "avantgarde"), 
         words = str_replace_all(words, "associationte", "associate"),
         words = str_replace_all(words, "atisk", "at-risk"),
         words = str_replace_all(words, "-support", "support"),
         words = str_replace_all(words, "-salary", "salary"),
         words = str_replace_all(words, "\\bmex\\.\\b", "mexico"),
         words = str_replace_all(words, "\\bcommis\\b", "commission"),
         words = str_replace_all(words, "\\bconsol\\b", "consolidation"),
         words = str_replace_all(words, "\\bcontribu\\b", "contribution"),
         words = str_replace_all(words, "renfection", "reinfection"),
         words = str_replace_all(words, "\\bpsy\\schiatry", "psychiatry"),
         words = str_replace_all(words, "nstitute", "institute"),
         words = str_replace_all(words, "associationtion", "association")) %>%
  filter(!(words %in% c("tory", "tional", "tural", "ture", "cal", "medi", "na", 
                        "cul","©", "v", "f", "r", "p", "e", "t", "l", "m", "n", 
                        "b", "m.s", "na", "fe", "ab", "db", "appbopbu", 
                        "apphopbia", "tionb", "tlona", "$", "mem", "ix", "viii",
                        "ogy", "sr.", "—a", "sss", "—lh", "—salary", "—support", 
                        "pp", "©rants", "mr", "mr.", "&m","à", "—dr", "ip", 
                        "ooo", "ps", "da", "b.sc", ".ft", "a.", "sr.", "a.b.",
                        "a.m.", "univ.", ".v.m", "xmv", "am", "cc", "du", "candi")))

tokenised.no.punct.nsw <- tokenised.no.punct.nsw %>%
  mutate(words = str_replace_all(words, "approappropriation", "appropriation"),
         words = str_replace_all(words, "approappropriations", "appropriations"),
         words = str_replace_all(words, "approapproriations", "appropriations"),
         words = str_replace_all(words, "adadministration", "administration"),
         words = str_replace_all(words, "engiricer", "engineer"),
         words = str_replace_all(words, "driveroffice", "driver office"),
         words = str_replace_all(words, "workerschallengebetween", "workers challenge between"),
         words = str_replace_all(words, "andworld", "and world"),
         words = str_replace_all(words, "whileeach", "while each"),
         words = str_replace_all(words, "joinother", "join others"),
         words = str_replace_all(words, "successfuldemostrations", "successful demostrations"),
         words = str_replace_all(words, "promoterenewal", "promote renewal"),
         words = str_replace_all(words, "dcnuclear", "dc nuclear"),
         words = str_replace_all(words, "initiativewhashington", "initiative whashington"),
         words = str_replace_all(words, "progressresolved", "progress resolved"),
         words = str_replace_all(words, "campaingfor", "campaingfor"),
         words = str_replace_all(words, "diseasesurveillance", 
                                 "disease surveillance"),
         words = str_replace_all(words, "networkschallengebecause", 
                                 "networks challenge because"),
         words = str_replace_all(words, "resilienceto", "resilience to"),
         words = str_replace_all(words, "forsmall", "for small"),
         words = str_replace_all(words, "impactinvesting", "impact investing"),
         words = str_replace_all(words, "industrychallengegovernment", 
                                 "industry challeng egovernment"),
         words = str_replace_all(words, "stewardshipjohn", "stewardship john"),
         words = str_replace_all(words, "foundationorgcondensed", "foundation org condensed"),
         words = str_replace_all(words, "activitiesdecember", "activities december"),
         words = str_replace_all(words, "rodinpresident", "rodin president"),
         words = str_replace_all(words, "presidentthe", "president the"),
         words = str_replace_all(words, "foundationorggrantsinescapable", 
                                 "foundation org grants inescapable"),
         words = str_replace_all(words, "withmore", "with more"),
         words = str_replace_all(words, "bostonhenry", "boston henry"),
         words = str_replace_all(words, "andthe", "and the"),
         words = str_replace_all(words, "countyhealth", "county health"),
         words = str_replace_all(words, "uponthe", "upon the"),
         words = str_replace_all(words, "wasreopened", "was reopened"),
         words = str_replace_all(words, "boardsrepresentative", 
                                 "boards representative"),
         words = str_replace_all(words, "thesanitary", "the sanitary"),
         words = str_replace_all(words, "weregranted", "were granted"),
         words = str_replace_all(words, "hadbeen", "had been"),
         words = str_replace_all(words, "variousprojects", "various projects"),
         words = str_replace_all(words, "servicesof", "service sof"),
         words = str_replace_all(words, "bythe", "by the"),
         words = str_replace_all(words, "waspermitted", "was permitted"),
         words = str_replace_all(words, "intoconsultation", "into consultation"),
         words = str_replace_all(words, "supplysystem", "supply system"),
         words = str_replace_all(words, "baltimoredonald", "baltimore donald"),
         words = str_replace_all(words, "adviserand", "adviser and"),
         words = str_replace_all(words, "yorkgeorge", "york george"),
         words = str_replace_all(words, "mainte nance", "maintenance"), 
         words = str_replace_all(words, "thevenereal", "the venereal"),
         words = str_replace_all(words, "iscorrelated", "is correlated"),
         words = str_replace_all(words, "epidemiologistwithin", "epidemiologist within"),
         words = str_replace_all(words, "tohonduras", "to honduras"),
         words = str_replace_all(words, "cooperativeprogram", "cooperative program"),
         words = str_replace_all(words, "madesurveys", "made surveys"),
         words = str_replace_all(words, "stationedin", "stationedin"),
         words = str_replace_all(words, "assistancein", "assistance in"),
         words = str_replace_all(words, "operationhookworm", "operation hookworm"),
         words = str_replace_all(words, "atechnical", "a technical"),
         words = str_replace_all(words, "amedical", "amedical"),
         words = str_replace_all(words, "otherplaces", "otherplaces"),
         words = str_replace_all(words, "undergovernmental", "undergovernmental"),
         words = str_replace_all(words, "hehas", "he has"),
         words = str_replace_all(words, "giveadvice", "give advice"),
         words = str_replace_all(words, "thecontrol", "the control"),
         words = str_replace_all(words, "statessanitary", "states sanitary"),
         words = str_replace_all(words, "malariastudies", "malaria studies"),
         words = str_replace_all(words, "servicesof", "services of"),
         words = str_replace_all(words, "ahonduran", "a honduran"),
         words = str_replace_all(words, "ingeneral", "in general"),
         words = str_replace_all(words, "bylending", "by lending"),
         words = str_replace_all(words, "fieldpersonnel", "field personnel"),
         words = str_replace_all(words, "physiciansone", "physicians one"),
         words = str_replace_all(words, "appropriationsa", "appropriations a"),
         words = str_replace_all(words, "healthwith", "health with"),
         words = str_replace_all(words, "underthe", "under the"),
         words = str_replace_all(words, "theservices", "the services"),
         words = str_replace_all(words, "localityin", "localityin"),
         words = str_replace_all(words, "inmany", "in many"),
         words = str_replace_all(words, "availablethe", "available the"),
         words = str_replace_all(words, "drainagesystem", "drainage system"),
         words = str_replace_all(words, "nationaldepartment", "national department"),
         words = str_replace_all(words, "localsanitary", "local sanitary"),
         words = str_replace_all(words, "themthe", "them the"),
         words = str_replace_all(words, "watersupply", "water supply"),
         words = str_replace_all(words, "departmentin", "department in"),
         words = str_replace_all(words, "foundationcontinued", "foundation continued"),
         words = str_replace_all(words, "malariologistand", "malariologist and"),
         words = str_replace_all(words, "theproblem", "the problem"),
         words = str_replace_all(words, "drainagethe", "drainage the"),
         words = str_replace_all(words, "adjacentuntiled", "adjacent untiled"),
         words = str_replace_all(words, "wasoffered", "was offered"),
         words = str_replace_all(words, "foundationssanitary", "foundations sanitary"),
         words = str_replace_all(words, "theunited", "the united"),
         words = str_replace_all(words, "expectedthat", "expected that"),
         words = str_replace_all(words, "hondurasearly", "honduras early"),
         words = str_replace_all(words, "hondurasa", "honduras a"),
         words = str_replace_all(words, "otherplaces", "other places"),
         words = str_replace_all(words, "undergovernmental", "under governmental"),
         words = str_replace_all(words, "servicein", "service in"),
         words = str_replace_all(words, "furnishedby", "furnished by"),
         words = str_replace_all(words, "malariastudies", "malaria studies"),
         words = str_replace_all(words, "controlin", "control in"),
         words = str_replace_all(words, "fordrainage", "for drainage"),
         words = str_replace_all(words, "furthertests", "further tests"),
         words = str_replace_all(words, "divisionin", "division in"),
         words = str_replace_all(words, "nicaraguacontinued", "nicaragua continued"),
         words = str_replace_all(words, "adivision", "a division"),
         words = str_replace_all(words, "governmenthaving", "government having"),
         words = str_replace_all(words, "ofthe ", "of the "),
         words = str_replace_all(words, "representativeserved", "representative served"),
         words = str_replace_all(words, "giveexpert", "give expert"),
         words = str_replace_all(words, "atwoyear", "a two year"),
         words = str_replace_all(words, "andthree", "and three"),
         words = str_replace_all(words, "latelybeen", "lately been"),
         words = str_replace_all(words, "mechanicalequipmentfillinga", 
                                 "mechanical equipment filling a"),
         words = str_replace_all(words, "breedingplace", "breeding place"),
         words = str_replace_all(words, "foundationfor", "foundation for"),
         words = str_replace_all(words, "prepareda", "prepared a"),
         words = str_replace_all(words, "beenlent", "been lent"),
         words = str_replace_all(words, "foundationto", "foundation to"),
         words = str_replace_all(words, "beingtrained", "being trained"),
         words = str_replace_all(words, "foundationfellowship", "foundation fellowship"),
         words = str_replace_all(words, "hondurasand", "honduras and"),
         words = str_replace_all(words, "installinga", "installing a"),
         words = str_replace_all(words, "themthe", "them the"),
         words = str_replace_all(words, "ofceylon", "of ceylon"),
         words = str_replace_all(words, "observationa", "observation a"),
         words = str_replace_all(words, "inthis", "in this"),
         words = str_replace_all(words, "furnishedby", "furnished by"),
         words = str_replace_all(words, "corporationswho", "corporations who"),
         words = str_replace_all(words, "infieldsketchingand", "in field sketching and"),
         words = str_replace_all(words, "sectionof", "section of"),
         words = str_replace_all(words, "maintainsheadquarters", "maintains headquarters"),
         words = str_replace_all(words, "aidedthis", "aided this"),
         words = str_replace_all(words, "asanitarytwoyear", "a sanitary two year"),
         words = str_replace_all(words, "bureauboredhole", "bureau bored hole"),
         words = str_replace_all(words, "departmentsbureauboredhole", 
                                 "departments bureau bored hole"),
         words = str_replace_all(words, "succesfullyinstalled", "succesfully installed"),
         words = str_replace_all(words, "villagesattention", "villages attention"),
         words = str_replace_all(words, "innorth", "in north"),
         words = str_replace_all(words, "practicableroutine", "practicable routine"),
         words = str_replace_all(words, "entomologistcollaborate", "entomologist collaborate"),
         words = str_replace_all(words, "foundationthe", "foundation the"),
         words = str_replace_all(words, "andof", "and of"),
         words = str_replace_all(words, "electroenceph alographyjohns", 
                                 "electroencephalography johns"),   
         words = str_replace_all(words, "europeedinburgh", "europe edinburgh"),
         words = str_replace_all(words, "institutionsgeorgia", "institutions georgia"),
         words = str_replace_all(words, "greecegermany", "greece germany"),
         words = str_replace_all(words, "fultonmontgomery", "fulton montgomery"),
         words = str_replace_all(words, "twonutritionists", "two nutritionists"),
         words = str_replace_all(words, "medicalschools", "medical schools"),
         words = str_replace_all(words, "therequest", "the request"),
         words = str_replace_all(words, "inbarranquilla", "in barranquilla"),
         words = str_replace_all(words, "anentomologist", "an entomologist"),
         words = str_replace_all(words, "staffmember", "staff member"),
         words = str_replace_all(words, "sanitarianthe", "sanitarian the"),
         words = str_replace_all(words, "providingsalaries", "providingsalaries"),
         words = str_replace_all(words, "commissionmade", "commission made"),
         words = str_replace_all(words, "intermittentlythe", "intermittently the"),
         words = str_replace_all(words, "medicalcommission", "medical commission"),
         words = str_replace_all(words, "factorsspecial", "factors special"),
         words = str_replace_all(words, "provedthat", "proved that"),
         words = str_replace_all(words, "thisregion", "this region"),
         words = str_replace_all(words, "wellestablished", "well established"),
         words = str_replace_all(words, "insuresnormal", "insures normal"),
         words = str_replace_all(words, "staffhas", "staff has"),
         words = str_replace_all(words, "antiguamontserrat vincentgrenadines luciadominica", 
                                 "antigua montserrat vincent grenadines lucia dominica"),
         words = str_replace_all(words, "incharge", "in charge"),
         words = str_replace_all(words, "ineurope", "in europe"),
         words = str_replace_all(words, "phaseof", "phase of"),
         words = str_replace_all(words, "researchphysicist", "research physicist"),
         words = str_replace_all(words, "withthe", "with the"),
         words = str_replace_all(words, "agriculturecomprised", "agriculture comprised"),
         words = str_replace_all(words, "anddoctor", "and doctor"),
         words = str_replace_all(words, "andsince", "and since"),
         words = str_replace_all(words, "auxiliaryworkers", "auxiliary workers"),
         words = str_replace_all(words, "subsequentobservation", "subsequent observation"),
         words = str_replace_all(words, "physiciansnurses", "physicians nurses"),
         words = str_replace_all(words, "equipmentgeorge", "equipment george"),
         words = str_replace_all(words, "worksengland", "works england"),
         words = str_replace_all(words, "draftsmen", "draft smen"),
         words = str_replace_all(words, "andsecretaries", "and secretaries"),
         words = str_replace_all(words, "yorkadditional", "york additional"),
         words = str_replace_all(words, "obstetricsand", "obstetrics and"),
         words = str_replace_all(words, "thetheatre", "the theatre"),
         words = str_replace_all(words, "museumlighting", "museum lighting"),
         words = str_replace_all(words, "governmentof", "government of"),
         words = str_replace_all(words, "screeningor", "screening or"),
         words = str_replace_all(words, "formfiltrationhas", "form filtration has"),
         words = str_replace_all(words, "recentbiological", "recent biological"),
         words = str_replace_all(words, "themathematician", "the mathematician"),
         words = str_replace_all(words, "analysisof", "analysis of"),
         words = str_replace_all(words, "lectromagneticfieldwhenan",
                                 "electromagnetic field when an"),
         words = str_replace_all(words, "meanta", "meant a"),
         words = str_replace_all(words, "peopledevelopment", "people development"),
         words = str_replace_all(words, "geneticistarthur", "geneticis tarthur"),
         words = str_replace_all(words, "agriculturewhich", "agriculture which"),
         words = str_replace_all(words, "themathematician", "the mathematician"),
         words = str_replace_all(words, "trainedprofessional", "trained professional"),
         words = str_replace_all(words, "foundationcolombiabogotacolombian", 
                                 "foundation colombia bogota colombian"),
         words = str_replace_all(words, "peopledevelopment", "people development"),
         words = str_replace_all(words, "foundationorgwhat", "foundation what"),
         words = str_replace_all(words, "foundationorgthese", "foundation these"),
         words = str_replace_all(words, "foundationrogduring", "foundation during"),
         words = str_replace_all(words, "theprovirice", "the province"),
         words = str_replace_all(words, "sanitaryengiricering", "sanitary engineering"),
         words = str_replace_all(words, "engiricerprovided", "engineer provided"),
         words = str_replace_all(words, "antimalariaengiricering", "antimalaria engineering"),
         words = str_replace_all(words, "engiriceringdepartment", "engineering department"),
         words = str_replace_all(words, "wellequippedlaboratory", "well equipped laboratory"),
         words = str_replace_all(words, "fundamentalstudies", "fundamental studies"),
         words = str_replace_all(words, "governmentfor", "government for"),
         words = str_replace_all(words, "presidentialdecree", "presidential decree"),
         words = str_replace_all(words, "humangenetics", "human genetics"),
         words = str_replace_all(words, "napleshas thefirstto", "naples has the first to"),
         words = str_replace_all(words, "governmentdevelopment", "government development"),
         words = str_replace_all(words, "environmentalsanitation", "environmental sanitation"),
         words = str_replace_all(words, "connectionwith", "connectio nwith"),
         words = str_replace_all(words, "increasingthe", "increasing the"),
         words = str_replace_all(words, "departmentsof", "departments of"),
         words = str_replace_all(words, "generalhealth", "general health"),
         words = str_replace_all(words, "anddynamic", "and dynamic"),
         words = str_replace_all(words, "asuniversity", "as university"),
         words = str_replace_all(words, "berlindahlem", "berlin dahlem"),
         words = str_replace_all(words, "andlandscape", "and landscape"),
         words = str_replace_all(words, "departmentof", "department of"),
         words = str_replace_all(words, "developmentof", "development of"),
         words = str_replace_all(words, "foundationsciencebased", "foundation science based"),
         words = str_replace_all(words, "atrisk", "at risk"),
         words = str_replace_all(words, "glycerinesalt", "glycerine salt"),
         words = str_replace_all(words, "budgetcontinued", "budget continued"),
         words = str_replace_all(words, "searchlight", "search light"),
         words = str_replace_all(words, "acquisitionloan", "acquisition loan"),
         words = str_replace_all(words, "massachusettsharvard", "massachusetts harvard"),
         words = str_replace_all(words, "medicalboard", "medical board"),
         words = str_replace_all(words, "ofa", "of a"),
         words = str_replace_all(words, "foundationhas", "foundation has"),
         words = str_replace_all(words, "developmgcountry", "developing-country"),
         words = str_replace_all(words, "contraceptivesand", "contraceptives and"),
         words = str_replace_all(words, "hereworkinprogress", "here work in progress"),
         words = str_replace_all(words, "kcontinued", "k-continued"),
         words = str_replace_all(words, "paymentsmissionary", "payment smissionary"),
         words = str_replace_all(words, "societieshos pitals", "societies hospitals"),
         words = str_replace_all(words, "ofinternational", "of international"),
         words = str_replace_all(words, "thcentury", "th century"),
         words = str_replace_all(words, "thecountry", "the country"),
         words = str_replace_all(words, "thegeneral", "the general"),
         words = str_replace_all(words, "thhealth", "the health"),
         words = str_replace_all(words, "activitiescontinued", "activities continued"),
         words = str_replace_all(words, "developmentcontinued", "development continued"),
         words = str_replace_all(words, "programscontinued", "programs continued"),
         words = str_replace_all(words, "organizationscontinued", "organizations continued"),
         words = str_replace_all(words, "institutionscontinued", "institutions continued"),
         words = str_replace_all(words, "healthcontinued", "health continued"),
         words = str_replace_all(words, "paymentsgeneral", "payments general"),
         words = str_replace_all(words, "payments research", "payments research"),
         words = str_replace_all(words, "dcontinued", "continued"),
         words = str_replace_all(words, "requirementscontinued", "requirements continued"),
         words = str_replace_all(words, "educationalcontinued", "educational continued"),
         words = str_replace_all(words, "schoolcontinued", "school continued"),
         words = str_replace_all(words, "budgetcontinued", "budget continued"),
         words = str_replace_all(words, "countriescontinued", "countries continued"),
         words = str_replace_all(words, "europecontinued", "europe continued"),
         words = str_replace_all(words, "statescontinued", "states continued"),
         words = str_replace_all(words, "paymentspublic", "payments public"),
         words = str_replace_all(words, "centerfor", "center for"),
         words = str_replace_all(words, "sciencebased", "science based"),
         words = str_replace_all(words, "agncultural", "agricultural"),
         words = str_replace_all(words, "costof", "cost of"),
         words = str_replace_all(words, "costsof", "costs of"),
         words = str_replace_all(words, "amencan", "american"),
         words = str_replace_all(words, "capacitybuilding", "capacity building"),
         words = str_replace_all(words, "civilsocietyorganizations", 
                                 "civil society organizations"),
         words = str_replace_all(words, "fouryear", "four year"),
         words = str_replace_all(words, "droughttolerant", "drought tolerant"),
         words = str_replace_all(words, "appbopria", "appropriations"),
         words = str_replace_all(words, "annourices", "announces"),
         words = str_replace_all(words, "developmentand", "development and"),
         words = str_replace_all(words, "foodand", "food and"),
         words = str_replace_all(words, "chma", "china"),
         words = str_replace_all(words, "erice$", "ence"),
         words = str_replace_all(words, "erices$", "ences"),
         words = str_replace_all(words, "engiricering", "engineering"),
         words = str_replace_all(words, "coricerning", "concerning"),
         words = str_replace_all(words, "(arice)(.*)", "ance\\2"),
         words = str_replace_all(words, "(erice)(.*)", "ence\\2"),
         words = str_replace_all(words, "(irice)(.*)", "ince\\2"),
         words = str_replace_all(words, "(nomi)ric(.*)", "\\1n\\2"),
         words = str_replace_all(words, "schoololarships", "scholarships"),
         words = str_replace_all(words, "artistinresidence", "artist in residence"),
         words = str_replace_all(words, "byaws", "by-laws"),
         words = str_replace_all(words, "newyorkcity", "new york city"),
         words = str_replace_all(words, "\\barc", "are"),
         words = str_replace_all(words, "afiveearperiod", "a five year period"),
         words = str_replace_all(words, "beview", "review"),
         words = str_replace_all(words, "composernesidence", "composer-in-residence"),
         words = str_replace_all(words, "inervice", "in-service"),
         words = str_replace_all(words, "\\bdesigna\\b", "designations"),
         words = str_replace_all(words, "berlinahlem", "berlin-dahlem"),
         words = str_replace_all(words, "andexplorations", "and explorations"),
         words = str_replace_all(words, "universityof", "university of"),
         words = str_replace_all(words, "ph.d.", "ph.d"),
         words = str_replace_all(words, "associa", "association"),
         words = str_replace_all(words, "\\bbabwe", "zimbabwe"),
         words = str_replace_all(words, "ogyscnpps", "scripps"),
         words = str_replace_all(words, "biotechnol\\b", "biotechnology"),
         words = str_replace_all(words, "sociol\\b", "sociology"),
         words = str_replace_all(words, "coricert", "concert"),
         words = str_replace_all(words, "ourwork", "our work"),
         words = str_replace_all(words, "arehaeology", "archaeology"),
         words = str_replace_all(words, "appbopbia", "appropriation"),
         words = str_replace_all(words, "researeh", "research"), 
         words = str_replace_all(words, "colomb\\b", "colombia"),
         words = str_replace_all(words, "minorityroup", "minority group"),
         words = str_replace_all(words, "artisthesidence", "artist in residence"),
         words = str_replace_all(words, "avantarde", "avantgarde"), 
         words = str_replace_all(words, "associationte", "associate"),
         words = str_replace_all(words, "atisk", "at-risk"),
         words = str_replace_all(words, "-support", "support"),
         words = str_replace_all(words, "-salary", "salary"),
         words = str_replace_all(words, "\\bmex\\.\\b", "mexico"),
         words = str_replace_all(words, "\\bcommis\\b", "commission"),
         words = str_replace_all(words, "\\bconsol\\b", "consolidation"),
         words = str_replace_all(words, "\\bcontribu\\b", "contribution"),
         words = str_replace_all(words, "renfection", "reinfection"),
         words = str_replace_all(words, "\\bpsy\\schiatry", "psychiatry"),
         words = str_replace_all(words, "nstitute", "institute"),
         words = str_replace_all(words, "associationtion", "association")) %>% 
  filter(!(words %in% c("tory", "tional", "tural", "ture", "cal", "medi", "na", 
                        "cul","©", "v", "f", "r", "p", "e", "t", "l", "m", "n", 
                        "b", "m.s", "na", "fe", "ab", "db", "appbopbu", 
                        "apphopbia", "tionb", "tlona", "$", "mem", "ix", "viii",
                        "sr.", "—a", "sss", "—lh", "—salary", "—support", 
                        "pp", "©rants", "mr", "mr.", "&m", "à", "—dr", "ip", 
                        "ooo", "ps", "da", "b.sc", ".ft", "a.", "sr.", "a.b.", 
                        "a.m.", "univ.", ".v.m", "xmv", "am", "cc", "du", "candi")))


# TODO quoi faire avec ce table? 
# table <- kableExtra::kable(head(tokenised.no.punct, 5)) %>%
#   kable_styling() %>%
#   add_header_above(c("Figure X: " = 2))

save(tokenised.punct, file = "tokenised.punct.Rda")
save(tokenised.no.punct.nsw, file = "tokenised.no.punct.nsw.Rda")
